generate_individual

simulates data for an individual participant

Usage

generate_individual(
  id = 1,
  T,
  P,
  sigma_residual,
  sigma_randint,
  main_rand = 0.5,
  rho,
  beta_logit,
  model,
  beta,
  theta1
)

Arguments

id

the assigned id for this specific individual

T

the number of observations for this individual

P

the number of measurements or predictors

sigma_residual

the noise added to the moderator formula when generate pseudo outcome. This noise is generated using ARIMA

sigma_randint

the noise added to the moderator formula so the intercept is random across participants when generate pseudo outcome.

main_rand

the noise added to the control formula when generate pseudo outcome. This noise is generated using ARIMA

rho

this value is used to simulate state or predictors and random noises. We also use ARIMA when generating predictors, and rho defines how much of the past values influence the current value in an autoregressive model

beta_logit

the true value of beta when simulate the probability of assigning treatment \(p(A_t = 1|H_t)\)

model

the true model for moderator when generate pseudo outcomes

beta

the true coefficients for moderator

theta1

the coefficients for the working model of control variables. For now, we just simply assume the working model contains the all predictors with coefficients equal to theta1

Value

A data frame containing the generated states \(S_t\), action \(A_t\), outcomes \(Y_{t+1}\), treatment effects, and noise components for a single participant. The outcome \(Y_{t + 1}\) is generated following the typical WCLS setting i.e. $$ Y_{i,t+1} \sim g_t(H_t) \theta_1 + (A_t - p_t(1|S_t)) (f(S_t)^T \beta + \sigma_{i,t,residual} + \sigma_{i,randint}) + \sigma_{i,t,rand} $$ where i represents the ith participant and t is the tth observation for this participant. This function simply assume the working model \(g_t(H_t)\) is \((H_{1,t} + H_{2,t} + ... + H_{P,t})\theta_1\)

Examples

generate_individual(id = 1, T = 50, P = 50, sigma_residual = 1.5, sigma_randint = 1.5, main_rand = 2, rho = 0.7,
beta_logit = c(-1, 1.6 * rep(1/50, 50)), model = ~ state1 + state2 + state3 + state4,
beta = matrix(c(-0.2, 0.8, 0.3, 0.7, 0.3),ncol = 1),
theta1 = 0.8
)
#>    id decision_point      state1      state2       state3     state4     state5
#> 1   1              1 -0.19765883  2.26788314 -4.585262503 -2.7211128 -3.7168160
#> 2   1              2 -0.88085245  3.13672917 -3.220530585 -0.9987536 -3.5482232
#> 3   1              3 -0.60846004  3.39440850 -3.916147950 -1.1938214 -4.0914216
#> 4   1              4  0.05705713  2.13855879 -4.055513504 -1.9502309 -2.9782735
#> 5   1              5 -1.23249505 -1.57975577 -3.042918746 -1.0873408 -2.0690745
#> 6   1              6 -1.21366124 -1.41837854 -3.424928860 -1.7552837 -4.4483785
#> 7   1              7 -1.75167731 -0.27818656 -2.221863098 -2.1892228 -5.3867214
#> 8   1              8 -4.01461005 -1.50704807  0.774858408 -1.8512215 -4.0373446
#> 9   1              9 -2.47802779 -1.83874816  3.635030225 -0.9776013 -1.0790549
#> 10  1             10  0.55975967 -1.44473233  3.253997526  0.7635798  1.8123920
#> 11  1             11  1.85020333 -1.88509062  2.322888389  0.9458231  2.1171852
#> 12  1             12  3.66787993 -3.38764364  2.756923391 -0.5758382  1.6711310
#> 13  1             13  5.88830681 -2.52691621  4.868324246 -3.3653061 -0.9843080
#> 14  1             14  6.23751528  0.55959851  5.572425315 -4.9205504 -2.7239699
#> 15  1             15  4.80612801  2.26047069  4.128911271 -3.9990856 -2.3172568
#> 16  1             16  2.76869897  2.93155173  2.733887869 -4.0931849 -2.3989163
#> 17  1             17  0.06503572  1.91105071  0.926859600 -5.3019524 -1.5995062
#> 18  1             18 -2.67392458  0.05131630  0.231689077 -4.0297899 -0.5077014
#> 19  1             19 -4.02614905 -0.51004602  0.821401552 -1.2897306 -1.9076132
#> 20  1             20 -4.72175172 -1.34376946  0.486393371 -0.4674784 -3.2624597
#> 21  1             21 -6.77930380 -0.84199936 -1.771058790 -0.8215639 -1.6938000
#> 22  1             22 -7.12922745 -0.43212901 -2.755554444  0.2375669 -0.3572353
#> 23  1             23 -4.16876616 -1.58282030 -0.567340615  1.1489274 -1.4288711
#> 24  1             24 -1.54209260 -0.12932953  0.001490199  0.3304749 -3.0343368
#> 25  1             25 -0.44814972  2.73817883 -0.182629212 -2.0115999 -2.2348640
#> 26  1             26  1.13311919  2.69417200  0.090369700 -1.7477318  1.6694068
#> 27  1             27  3.34509376  1.60498623  1.133277465 -1.5123362  2.0074240
#> 28  1             28  3.58367200  0.67014175  2.533743419 -3.7199404  0.6334232
#> 29  1             29  3.55654852  0.70215484  0.981912249 -4.6466137  1.3542778
#> 30  1             30  2.78334502  3.31400284  0.141182650 -3.9439135  0.5776489
#> 31  1             31 -0.36760173  3.65814501 -0.992883294 -3.0918822  2.3801420
#> 32  1             32 -3.65914690  2.17234604 -0.346951287 -2.7974285  3.9082142
#> 33  1             33 -6.07399640  1.50998300  1.688766043 -2.3951340  2.0358020
#> 34  1             34 -5.92557993  1.62259270  1.140181830 -1.0396000 -1.5723116
#> 35  1             35 -5.69514952  1.37930251  1.118095684  1.3749734 -2.4294371
#> 36  1             36 -3.04656879 -0.01010014  1.029155927  2.6408575  0.3845477
#> 37  1             37 -0.59888913 -0.29561071  1.024291493  0.4027388  0.7340721
#> 38  1             38 -3.24826384 -2.03937753 -0.275977920 -2.5666518 -0.3418253
#> 39  1             39 -4.31724439 -2.95806510 -1.627678404 -3.4298073 -2.0185226
#> 40  1             40 -5.44041102 -1.44384075 -0.976566989 -1.6869659 -1.4694064
#> 41  1             41 -6.73891016 -2.05606591 -0.874541865 -0.3359987  1.5948400
#> 42  1             42 -4.28295394 -3.11162653 -0.680274775  0.7667642  2.1513665
#> 43  1             43 -2.16928058 -1.96071491  0.771418978  0.2241800  1.2843466
#> 44  1             44 -3.71689601  0.23059453  1.260535207 -1.0831201  0.2104782
#> 45  1             45 -5.54407928 -0.69603613  1.437526266  0.3427186  1.2661508
#> 46  1             46 -5.06997765 -2.62357323  2.026378085  0.4768496  3.1159091
#> 47  1             47 -3.96352534 -4.04893364  1.739842484 -0.0545017  1.3000872
#> 48  1             48 -1.79404807 -5.84498433  1.174843819  0.7908770  1.5082834
#> 49  1             49  1.96857122 -6.41207954 -1.719194028  1.6595358  2.5621724
#> 50  1             50  4.60067925 -5.42731492 -3.915735070  0.4681038  0.3875824
#>        state6      state7     state8     state9    state10     state11
#> 1   0.6550341  2.26264454 -1.3935149 -3.9364436 -4.8017927  3.50845523
#> 2   0.9708717  2.90091199 -2.2327195 -3.5658871 -1.5691588  6.59632737
#> 3   1.6898336  4.53083936  0.2066985 -2.1913117 -1.3812709  6.73773973
#> 4   2.8140331  3.55787237  1.0754748  1.1692930 -2.3215395  2.27695839
#> 5   1.2099298  0.46051334  0.6681178 -0.1143020 -3.4698135 -1.34453956
#> 6  -0.6712339  1.02589062  0.4827324 -2.2168479 -3.8843906 -2.56630247
#> 7  -0.2805822  0.42042429 -0.2207173 -2.0885648 -3.2810954 -0.88798179
#> 8  -0.5387000  0.48848876  0.4754533 -0.5258551 -0.3048366  1.11684750
#> 9  -0.7706378  2.86446230  1.2936941  1.0624459  1.7615722 -0.59230632
#> 10 -1.0865318  2.18041419  1.3546707  1.9961373  3.1305735 -1.40966588
#> 11 -1.1115948  2.25345836  1.1052898  2.1501088  1.6939749  0.11278431
#> 12 -1.0815591  4.48484290  1.1737874  0.3361850 -1.2301605  0.55034700
#> 13 -0.4705849  4.40060373  2.1716159 -1.4515613 -2.1671048 -0.15992324
#> 14  1.0582973  3.43758167  2.7160768 -1.2945239 -2.9231939  0.75797201
#> 15 -0.5146646  2.30815758  3.2743696 -0.2430754 -0.9853225 -0.12704422
#> 16 -1.7895022  2.97505140  3.7494645 -0.5960822  1.0814649 -0.78152118
#> 17 -1.7253402  1.90813983  2.6385374 -1.4131699  0.8929583  1.41286865
#> 18 -2.9922502 -0.03714929  1.8402899 -0.9125111 -0.5746813  1.26538037
#> 19 -0.7515273  1.55480466  1.6055766  0.1936569  0.3570392 -0.46921007
#> 20  3.0173453  1.45931792  1.0940147 -0.2685811  3.2984303 -1.65691626
#> 21  4.0154883  0.42269341 -1.9005658  0.5576978  4.8718449 -1.81780886
#> 22  2.1513951 -0.64639502 -3.3612017  0.5870503  3.0855493 -0.02921163
#> 23  1.7413855 -2.21871435 -1.0549257 -1.6035350 -0.6111991  3.12877216
#> 24  2.9599005 -3.15881358 -1.9379533 -0.8620808 -1.6739655  3.59515278
#> 25  2.4292112 -4.05562267 -4.1943923  0.1290209 -0.2910391  0.85982797
#> 26  0.7297204 -4.31199814 -3.3817983 -1.6809055  0.5473328 -0.73344244
#> 27 -1.9399742 -3.96977698 -0.9445575 -0.8787931  0.8446156 -0.06763016
#> 28 -2.0352223 -3.87569668  2.3425026  1.5749841  2.2459272  1.30937902
#> 29 -0.6496191 -4.11358831  3.9208338  1.3773131  3.1423036  1.83327034
#> 30  1.0558541 -5.14755582  2.9435539 -0.7529664  3.5162277  3.25569830
#> 31  3.2527797 -5.51351998  2.9435005 -0.4088949  2.1898689  6.39353330
#> 32  3.3589835 -5.02976168  4.2191974  0.9096115  1.1898245  7.70153395
#> 33  2.0312208 -5.32724187  5.1876067 -1.6174193  1.2033267  6.14280885
#> 34 -0.1209388 -5.07613319  6.1039978 -2.5099068  1.1194867  3.81511567
#> 35 -0.8855927 -4.68477213  8.4139033 -1.5633457  1.7039324  3.55981642
#> 36 -1.3057367 -1.65254973  9.4893783 -0.7247795  2.7858996  3.96146137
#> 37 -4.3510648  2.17153435  7.9230414 -0.5588380  4.6592843  2.34478073
#> 38 -4.6461533  1.86970109  6.8028004 -0.8311110  5.0668650  1.09790003
#> 39 -2.8266390 -0.17382779  6.6496292 -1.0386003  4.7962435  1.44358827
#> 40 -2.6159673 -0.75822759  7.8202814 -0.3606056  3.5899032  1.74678179
#> 41 -1.4535788 -0.51297683  6.8294340  0.5194514  1.0898680  0.68064049
#> 42  0.3640707  1.03319655  4.3177800 -0.5924252  0.7351477 -0.74040645
#> 43 -0.7312373  2.87285054  4.5108329 -1.0261410 -0.5065523 -0.83492126
#> 44 -0.5047310  4.52539594  4.0046060 -1.3856799 -2.6176687  0.64832709
#> 45 -0.2586961  5.53488741  3.6441683 -2.0229858  0.7413933  1.42573276
#> 46 -1.7833538  5.31304041  5.0396187 -1.9023999  4.0995967  0.48812577
#> 47 -0.2716409  4.53492488  6.5540071 -1.2518679  1.9061580 -1.26068625
#> 48  2.0478093  0.93792909  6.5615807 -2.2663332 -1.7879117 -1.87512897
#> 49  3.5796174 -0.66153381  3.9297657 -1.8963225 -2.1156162 -1.87674768
#> 50  2.0839459  0.55644151  3.3937035  0.7735826 -1.8910171 -2.95833100
#>        state12    state13     state14     state15     state16    state17
#> 1  -0.05469091  0.3015784 -2.30647972 -4.64924921  0.45122367  1.7467729
#> 2  -0.73787888  0.6708354 -0.08289296 -3.75378728  0.94003762  3.2062272
#> 3  -2.63619238  0.4852857  1.88272405 -2.12457398  0.49041488  3.9631832
#> 4  -6.41126684  0.6769206  1.94333109 -1.21097179  0.92876144  3.5004086
#> 5  -8.86134801  2.5683346  1.88977626 -2.67471795  2.22178911  3.1174133
#> 6  -8.59844677  4.0083253  1.25798110 -4.56690245  6.02691084  4.6480904
#> 7  -6.01230858  1.7647057 -0.74718427 -5.04533301  5.88492763  5.3416195
#> 8  -2.60495175 -0.9215662 -2.59649618 -3.26123784  2.11064623  4.1559917
#> 9  -2.19747961 -1.1077442 -2.94012632 -2.12600029  1.63333833  2.3582297
#> 10 -1.18076955 -0.1873512 -2.32264028 -2.47280009  1.74888293  2.6810132
#> 11 -0.06059714 -1.2152418 -3.10838029 -0.23496456  0.53611257  4.0105879
#> 12 -0.46471579 -2.7332790 -4.06995842  0.59040505  0.91822688  3.9864155
#> 13  0.02797149 -2.2446383 -2.64552185 -1.34945288  1.49152572  2.2130081
#> 14 -0.35705550 -0.2597581 -3.26603484 -1.48486324  2.59450825  0.7556880
#> 15 -0.68088401  1.2352552 -3.86610573 -0.49865077  4.73892978 -0.1564006
#> 16  0.49468857  1.8268860 -1.71891696  0.54243283  5.10629184 -1.7891947
#> 17 -0.42577097  2.7305987 -1.17704175  1.18013255  5.12148721 -3.2148466
#> 18 -1.56605501  3.4491560 -1.14191748 -0.66039065  3.81887259 -3.8105856
#> 19 -2.85683077  2.6788539 -2.85629924 -2.27692447  2.35253958 -2.8819347
#> 20 -3.32613202  1.9147407 -5.36498446 -2.11900263  1.15372256 -2.2974040
#> 21 -1.25549541  1.1615636 -5.81997802 -1.02389792 -0.59236930 -2.7514663
#> 22  1.58816044  0.3622439 -4.94578392  0.82998852 -1.26863038 -3.3673301
#> 23  1.98739610 -0.9118171 -2.36722854  1.81763079 -3.51358200 -3.5781636
#> 24  1.47103639 -0.9558217  0.12911617  0.85394291 -4.02144195 -2.5438977
#> 25  3.38282627 -0.5311123  0.21206059  0.17459984 -2.09139286 -2.6759193
#> 26  4.93290852 -1.9658200  0.48257003  0.93690654 -1.54657712 -3.0828985
#> 27  3.48769966 -2.1075134  0.82866208  1.25958760 -0.95185691 -1.5268952
#> 28  1.48833398  0.2056480 -1.37564831  2.18306805 -0.27956542  0.3780160
#> 29  0.95575955  2.6156982 -2.23564759  4.94914236 -1.99140101  2.1816521
#> 30  0.86537819  2.4619242 -1.53445498  5.91055849 -2.20839491  3.1796245
#> 31 -0.52744319  2.8777229 -1.43776493  5.07879970 -0.09971761  4.0788549
#> 32 -0.70640182  2.3466715 -1.93342184  4.28996322  1.60560159  3.1684000
#> 33 -0.89408440  1.6401257 -1.75033171  4.30371423  3.98085117 -0.1713574
#> 34 -2.73469507  3.0285322 -0.96779765  3.08070249  3.53121985 -2.7500841
#> 35 -2.81552762  3.3798610  0.10367386  2.05844681  1.45245969 -5.0813993
#> 36 -3.03500362  1.5095525 -2.51459796  2.60606042  1.02993986 -7.0548771
#> 37 -1.84939322  0.3698769 -4.47488520  3.05739699  1.95323998 -6.2865960
#> 38 -0.27525534 -0.3412872 -3.35484763  2.52737726  3.71889809 -3.5944408
#> 39 -0.71658796 -1.7108768 -1.40435527  0.84989470  5.02687374 -2.0392006
#> 40 -0.37614741 -2.2215299 -0.02563802 -0.32775382  4.84597544 -1.6698656
#> 41  1.67681774 -2.4225334 -0.75482680 -0.31112809  2.64333200 -1.2963683
#> 42  3.23104841 -2.2990898  0.15399506 -0.43796005  1.97827193 -1.2292233
#> 43  2.69272811 -2.4026356  1.20622812 -2.53418964  1.80401928 -0.9503100
#> 44  1.79729088 -4.3092378  2.78814543 -3.33327963  2.20726478 -1.4836391
#> 45  1.42339581 -4.5578207  3.28956213 -1.77806789  2.64576210 -2.0216713
#> 46  1.87871306 -4.0779895  2.62219311  0.25852905  0.91050731 -1.5954484
#> 47  2.62847437 -3.7809251  2.56718888  0.03827131  0.31017378 -2.1367686
#> 48  2.02870854 -2.9530178  1.96276331 -1.59328818 -0.26234404 -2.1366509
#> 49  0.94001925 -2.2277944  0.38391352 -1.81726213 -1.32728564  0.2924049
#> 50 -0.18849366 -2.6535705 -2.36732423 -0.41555547 -2.36419950  2.2998561
#>        state18      state19     state20    state21     state22     state23
#> 1   0.19499927  0.358097268  3.08131847 -2.1168132  1.83571929  0.50620948
#> 2  -2.05108400  0.995139466  2.39309096 -2.2997767 -1.87745079 -0.84565056
#> 3  -2.47450185 -1.788050832  2.29500894 -1.4455096 -1.11515047 -0.28775691
#> 4  -3.41645687 -4.077974595  1.99702667  1.9672443  0.71820263 -1.03249028
#> 5  -2.16982861 -2.079775351  1.76871080  3.6208087 -0.47114627 -1.40018422
#> 6   0.18288981  0.121504676  0.91058546  1.7905346 -2.62699216 -2.61240655
#> 7   0.29485353 -1.319594219 -0.09785110 -2.0164040 -3.17472387 -3.71949431
#> 8   0.46662229 -3.193608470  0.32986716 -1.9297474 -2.54742408 -1.95853967
#> 9   0.13073435 -2.914552550  0.58884435 -1.2196368 -3.24339287 -0.59148126
#> 10  0.70661713 -1.722275768  1.27086562 -2.6304738 -4.45963726 -1.75389261
#> 11  2.57624873 -1.154809786  2.53378442 -1.5559183 -3.60857865 -3.34523976
#> 12  4.00613359 -1.879114550  4.46939810 -0.9030719 -2.62389408 -1.84168608
#> 13  3.95155706  0.494651849  4.34506886 -0.5456403 -3.76104580 -0.69890935
#> 14  1.64294407  3.858061709  0.68958566 -0.6706959 -4.64329973 -0.74752147
#> 15 -0.01491314  3.300432381  0.05247086 -0.4744316 -3.89901190  0.25193477
#> 16 -0.21016818  2.658806281  0.23448821 -0.7842577 -2.58360735  0.50424385
#> 17 -0.51298061  1.174375466 -1.19789725 -2.4735560 -0.74796370 -0.41407771
#> 18  2.22683348  1.340392094 -0.03515300 -0.4713846  0.45572203 -0.10379809
#> 19  3.44825551  2.671274493  1.30634403  0.2534051  0.53857678 -0.02469221
#> 20  0.77387706  2.514841796  2.15374139 -3.3337786  1.92282693 -0.34054026
#> 21  0.11259087  2.590102544  3.56079409 -5.2493170  0.50266808 -0.10317435
#> 22  1.67586010  2.019398368  5.29345982 -4.9835630 -3.71682172 -1.57703536
#> 23  2.50192523  0.927977953  6.95192641 -4.8500072 -5.53028355 -1.00922430
#> 24  0.37653699  0.353714873  5.40424508 -4.0899778 -4.99226264 -1.00880159
#> 25 -2.23182652 -1.060365660  2.63882675 -4.4720947 -3.17961914 -2.48422076
#> 26 -2.52085473 -2.432136439  1.73044274 -4.5306503 -2.09288962 -2.00574616
#> 27 -1.66739156 -2.332020324  1.75388463 -2.9495431 -1.27002217 -2.60475745
#> 28 -0.17479914 -0.820890098  1.31097517 -1.6827330  0.35099251 -3.54250500
#> 29 -0.13077981  0.497576464  1.65427714 -0.2300945  3.02109452 -4.51252020
#> 30 -1.96179394 -1.858635268  3.27645515 -0.6333868  3.44750229 -5.82691442
#> 31 -2.95873570 -3.639752114  3.40336466  0.9214698  2.51922973 -6.47947910
#> 32 -2.80758435 -3.094565005  1.46283055  5.2527345  2.22015779 -5.88400991
#> 33 -3.60095587 -1.077347448 -1.02385278  8.2009003  1.32775708 -4.64419352
#> 34 -4.36002740  0.468374426 -1.40858936  8.6010470  2.03520998 -4.00633316
#> 35 -3.29038739 -0.158974143 -0.25572376  6.5660281  3.05479104 -5.54788181
#> 36 -2.86817066 -1.614194473  0.04956268  2.9037265  2.68921575 -6.73292461
#> 37 -2.86056438 -2.321154486 -1.18485879  0.5387252  0.99567684 -5.21222475
#> 38 -3.51552003 -0.927053610 -2.68698629  0.9024530 -0.46016086 -3.70565169
#> 39 -4.81097066 -0.387387006 -4.06888927  1.9636787 -1.12208143 -1.13766112
#> 40 -5.69044903  0.003374311 -3.95706083  4.2213930 -2.17649358  0.72292615
#> 41 -4.94703973  0.185222923 -3.97087124  4.9031082 -4.42301736 -0.05343848
#> 42 -2.65171355 -1.145696625 -4.44883091  4.4439047 -6.23368075 -1.00987216
#> 43 -0.42781415 -2.567587685 -3.14315086  4.6535623 -5.41461170 -0.93978329
#> 44  1.22016856 -2.287281866 -1.54426000  2.8855041 -4.40100646  0.02162884
#> 45  0.62098674 -0.171377974 -1.37678282  1.0016272 -2.89813255  0.26185417
#> 46 -0.54236753  0.325592493 -3.14818784  0.1780170  0.04755593  0.77226599
#> 47  0.05347651  0.287204656 -3.41259104  0.8694256  1.18473304 -0.24599927
#> 48  0.96469957  1.089371351 -3.82557542  1.4537190  2.09147327 -2.23662847
#> 49  0.26138977  0.666438436 -3.88054245  1.1362773  2.74026944 -2.06257928
#> 50 -0.38191918  1.038366248 -3.43739767  0.8643780  2.30688909 -2.20616987
#>       state24    state25     state26      state27     state28     state29
#> 1  -4.7683614  0.8987155 -2.16375412 -0.363062654  0.57356946  1.29738830
#> 2  -2.8912040  0.3792985 -1.60215182  0.642357904 -0.10234915  1.72980389
#> 3   0.2207401  1.9080754  0.36318940  0.670070658 -0.84717161  1.27369919
#> 4  -0.4241693  1.7016292  1.56360779  0.775646156  0.07055651  0.51570368
#> 5  -1.8574688 -0.9037049  2.96545506 -0.238441195  2.64299248 -1.05523092
#> 6  -2.6547383 -3.7249977  3.17845530 -1.627805411  3.20305449 -4.05720771
#> 7  -2.8613751 -5.9059816  2.59732956 -1.356470523  2.20804224 -5.61009900
#> 8  -1.5412959 -7.1698576 -0.01622275  0.209722039  3.84945864 -5.14860217
#> 9  -0.3857932 -5.7749231 -1.31694878  0.156267077  4.40363594 -5.02032117
#> 10  2.6159872 -5.8522468 -0.86510317  0.699334413  4.59766526 -4.86380080
#> 11  4.0759823 -5.7025585  0.39965139  3.050710573  6.31808926 -5.37836856
#> 12  2.3104402 -3.1831027 -0.07467211  5.679617098  6.79952330 -3.52516790
#> 13  0.6100724 -4.0534569 -1.39479994  6.000624748  6.80732451  1.49852745
#> 14  0.8061338 -4.2129758 -0.99054447  4.858148229  5.68500530  2.77820426
#> 15  1.3664644 -2.1041584 -1.12499145  3.454119420  4.04462796  2.54014189
#> 16  2.8722489  1.1120079 -2.01112239  1.972086827  3.48918880  2.38228010
#> 17  4.0120117  2.7675666 -1.53442395  0.139910657  4.24940197  1.90292375
#> 18  3.6530295  1.0657972 -0.53029174 -0.425690852  5.93306993  2.84329666
#> 19  4.3298613  1.0187822 -0.65006830  0.810813648  4.12258479  1.10018339
#> 20  2.0353253  1.2866373 -0.49139089  1.450508936  1.29024655  0.37011576
#> 21 -1.4843871 -0.2303428 -0.23384597  2.288077861  3.00509782  0.19566935
#> 22 -1.9219360 -0.5708499 -1.14109318  3.056699196  3.21908938 -0.09828335
#> 23 -1.0094736  1.5294365 -0.61732570  2.552616563  1.06800916  0.69380114
#> 24  0.4058834  2.5891412  1.71150014  1.384466745 -0.35085310 -0.40668082
#> 25  2.6712400  1.9201383  2.49449651  0.632703794 -0.85938848  0.08108718
#> 26  2.6782066  1.2137211  2.31035581  0.002436192 -0.92617502  1.44949046
#> 27  1.1543490  2.2397884  1.84759740  0.530986433 -0.46287930  3.31334728
#> 28 -0.6447449  3.7013420  0.89962435 -0.756503223 -1.17986898  3.33030801
#> 29  0.5314817  4.5007070  1.24424057 -0.914401673 -3.80971545  0.74788799
#> 30  1.9300452  2.0869735  1.61166459  0.925939085 -2.84726100  0.46546235
#> 31  2.2173355 -2.6418538  1.77409275  2.173161856  0.62454050 -0.02200456
#> 32  4.2361087 -3.4202262  2.08271399  3.346568532  0.57010922 -1.48789251
#> 33  3.5956429 -0.6966130  2.35045187  2.817755139 -0.82322272 -1.82203425
#> 34  0.8815383 -0.5535705  1.88500156  1.777649274  0.64402871 -1.94764066
#> 35  0.9708597 -3.2670651 -0.77602555  0.770722970  3.69300369  0.75963024
#> 36  3.4193568 -1.9764905 -1.38997262 -0.930192714  5.80178106  1.80702911
#> 37  5.0118933  0.1724987 -1.49764754 -0.523735151  5.71187851 -0.33857919
#> 38  5.2721419 -0.7185346 -0.72894620 -0.575269168  3.40233057 -1.17596315
#> 39  3.0784854 -1.5762988  1.45059083 -1.343249397 -0.05776577 -1.64302189
#> 40  1.1010824 -1.2009392  1.26791889 -1.563345057 -0.63611746 -0.87333232
#> 41  1.1121647  0.5256319  1.51529604 -2.020821789  0.09924616 -2.25603247
#> 42  0.5379826  0.5798794  1.50573503 -1.202883483  0.46412621 -3.90100123
#> 43  1.4228230 -0.9662800  0.14418662 -0.134507464  1.43568682 -2.42230442
#> 44  2.5227551 -0.3341582 -1.28824179  0.664546781  2.53146458 -1.12599486
#> 45  2.3466909  2.1343431 -0.91839478 -0.845787536  3.36878665 -0.96599136
#> 46  3.5062581  2.9872497  0.30562203 -2.266008759  2.67112775 -0.87933199
#> 47  3.6465324  0.7730803  0.55073802 -2.322333660  2.52155045  1.60472102
#> 48  2.1063361 -2.2938108  2.09463881 -0.407553658  3.08439382  3.40109620
#> 49  0.9084975 -2.2587872  3.25409931  1.492772690  2.63453963  2.05375752
#> 50 -0.6772878 -0.1551558  2.35215404  0.415654838  1.02707277  0.75922234
#>        state30      state31    state32    state33     state34      state35
#> 1  -0.19763670 -2.293478701  3.3965041 -1.1208551 -0.89226607 -0.296103953
#> 2   1.08287732  0.160822756  2.7280979 -2.2808679 -1.76694674 -0.962911374
#> 3  -0.07349476  1.262503428  0.9062102 -1.4539729  0.55517400 -0.229342933
#> 4  -2.35293675  0.641592188  0.8824735 -1.5399376  2.12207885  0.663650116
#> 5  -3.96154939 -1.733802745  1.3305504 -0.6053096  0.88803632  1.253166434
#> 6  -4.07599823 -3.469132793  0.6342425 -0.3413106  0.46757424  2.928521406
#> 7  -4.67172449 -2.646279503 -0.9867930 -0.4722614  2.03931801  0.812475540
#> 8  -4.95439705 -2.395298558 -1.2299511  1.5470136  2.61267240 -0.243176630
#> 9  -4.21268125 -0.304304812 -0.3457003  2.9186470  0.26452607  1.829782437
#> 10 -4.33785089  3.519118395 -0.6330503  2.9896248 -2.17101955  2.989457418
#> 11 -6.71781321  3.403407026 -1.3744438  2.7195556 -3.05066675  1.457920708
#> 12 -8.48824289  0.913055496 -1.8664681  4.1264103 -3.30265388 -1.537661668
#> 13 -7.82218295  0.710640739 -1.5197467  6.6724437 -2.99523748 -3.485826210
#> 14 -4.34346105  1.669167645 -0.3030892  6.8851858 -3.44518409 -3.039469105
#> 15 -2.78455200  0.814014246  0.5236319  4.7975255 -3.17973172 -1.412492013
#> 16 -3.67473740 -1.057168798  0.3851173  2.3665182 -2.68870529 -2.022181281
#> 17 -1.34906789 -1.287366112 -0.7204474  1.9098494 -3.03110785  0.282069743
#> 18 -1.29997390 -2.618497549 -2.2808129  1.4277904 -3.64774499  2.304168457
#> 19 -2.21706376 -4.943952162 -3.3419904  0.2529964 -4.54330251  0.750648805
#> 20 -1.20268447 -4.986365832 -3.7208542  1.8436990 -3.58020983  0.468667788
#> 21 -3.18137312 -3.347849129 -3.1404382  5.6265067 -1.77231420  2.202632705
#> 22 -4.16388919 -1.173755269 -4.1154798  7.4974352  0.06198753  2.301097613
#> 23 -2.79796966  0.129099855 -3.8446487  6.6599397  3.33987337  2.584326398
#> 24 -0.71364488 -0.859547928 -0.7321189  4.8879640  6.06347251  3.989202446
#> 25  0.32323822 -2.620396172  1.6230741  1.0021775  5.06721084  4.113106540
#> 26 -0.35808901 -3.731287830  1.1378055 -1.8868265  2.34304016  3.915536440
#> 27 -1.31982061 -2.749552479 -1.3530393 -1.5158040 -0.73276388  3.293648375
#> 28 -3.23326458 -1.890034397 -3.0359748 -1.4634682 -2.13205076  1.835902016
#> 29 -3.57309078 -1.761633623 -5.1654531 -2.6412233 -4.27943207 -0.693528148
#> 30 -1.58484392 -0.498959513 -5.8567534 -3.0011328 -5.52881112 -0.956914715
#> 31 -0.75149792 -2.187067493 -3.9711713 -4.9755982 -2.72858645  1.345235248
#> 32 -0.09045863 -3.928198206 -1.7613317 -7.2424432 -0.70985533  2.855981214
#> 33  0.71219732 -1.402404184  0.2619336 -6.6605706  0.17163350  1.506205430
#> 34 -0.27596284  2.106632997  0.4846843 -5.3309587 -0.72053940  0.161470164
#> 35 -1.35474542  3.003240154 -1.8691576 -4.7010229 -2.26234348  0.001687402
#> 36 -2.01028986  3.292206651 -2.7942488 -3.0563918 -1.40364543 -1.100604695
#> 37 -2.80547860  3.593161931 -0.9632207 -1.4012102  0.34885076 -2.966312902
#> 38 -4.06016012  2.080346541  1.0127439 -2.1900878 -1.34370425 -3.749097945
#> 39 -4.55541426  1.471757042  0.9767442 -1.6818600 -2.79373668 -1.486047600
#> 40 -2.35814058  0.578970102 -0.1812501 -1.1448396 -3.44305417 -0.009784431
#> 41  0.82774839  0.347470201 -0.2284318 -1.4494152 -3.45262530 -1.252547428
#> 42  0.47322824  0.008271869  1.6888330  0.2137450 -2.67583421 -1.774280111
#> 43 -0.82931071  0.305908364  0.6672258  0.6316305 -1.74628675 -2.783548318
#> 44 -2.36897353  3.336076856 -1.3474208 -0.8096455  0.39233475 -2.622969715
#> 45 -4.37199964  5.279183967 -1.4080744 -2.1288101  1.41413675 -0.542885986
#> 46 -2.66010880  4.017306352 -1.5969491 -3.9153022  2.54106297  0.495338629
#> 47 -2.23820960  3.107788380 -1.1765019 -4.0096689  2.75421922 -0.185417822
#> 48 -2.55634234  2.005030852 -0.1883766 -2.6697599  2.56370661 -1.739810969
#> 49 -2.78112317 -0.533137421  0.4505637 -1.0191058  2.42382317 -2.330398757
#> 50 -2.47438152 -1.146798353 -1.7798297  1.8071713  1.19821286 -3.413229755
#>          state36     state37     state38     state39     state40      state41
#> 1   4.5318092103  3.09723080  0.92873069  1.19248029 -0.85877898  2.116510400
#> 2   5.2057912456  1.65252668  0.35350847  2.30364728  0.26038453  1.406382374
#> 3   3.0062928766 -0.19446154  2.78601340  2.20383470 -0.08443651 -0.844181517
#> 4   2.1582792767 -0.36239500  3.51132574  1.13620454 -0.11380223 -2.903202972
#> 5   3.8245136367  0.14731191  2.49759706  0.04392027  0.41340187 -2.297194064
#> 6   4.8299370924  0.72093555  3.23957289 -0.76951247  1.10563672 -1.414468223
#> 7   5.1300005246  1.04429864  2.36924910  1.16043107  0.55703883 -0.898526965
#> 8   3.2035720977  0.94003875  0.31036216  0.68076352  1.72102191 -0.132604833
#> 9   1.0109079061  0.64190464 -0.47022324 -1.96743205  3.77179401  1.006637511
#> 10  0.2930257228  0.39325523 -1.58273232 -0.65011767  2.34471203  1.319808087
#> 11 -0.0003217831 -0.98377883 -1.59940351  0.78296933  0.69614266  2.544920676
#> 12 -0.2045608787 -1.53426849 -0.54496525  1.77742281 -0.72981370  4.549361744
#> 13 -2.4972400128  0.01529885 -0.36557435  2.00779235 -1.51706244  2.873000750
#> 14 -3.6365158953  0.95315188  0.03901117  1.34544370 -0.68508769 -0.057720511
#> 15 -3.6992931178  0.24659996  1.18075231  0.94900134  1.12088642 -2.590947222
#> 16 -4.8723276255 -0.57793040  3.35007920 -1.68463294  1.61922372 -4.577828281
#> 17 -5.2822349716 -0.25882771  4.37700646 -5.60568846  0.85041631 -3.049666889
#> 18 -3.0122967142  1.04435813  2.55659878 -4.38036383  1.01402398 -0.141961991
#> 19 -0.3771547663 -0.64411924  2.40392807 -0.11281401  2.21744194  2.330042690
#> 20 -0.8282120357 -2.51893670  2.40662208  2.99936252  1.91118507  2.257308286
#> 21 -1.7590494300  0.32232067  1.41107367  1.97797963  1.34650941 -0.348607143
#> 22 -1.5970864366  0.19940565  1.19039496  0.19636228  3.04964001 -2.909562481
#> 23 -1.5098531743 -2.43259268  0.69469829  1.24766998  4.05751402 -3.971079246
#> 24 -0.7876593468 -2.11463970  0.17763105  1.04498904  4.04759050 -2.548230693
#> 25  1.3164759650 -0.09456975  0.24326557 -1.13027637  4.24881008 -1.308480616
#> 26  3.0869607356  0.68991913  1.36500050 -0.69043672  4.97343664 -0.909006555
#> 27  1.3278818474  1.42177360  2.23525430  2.41764607  4.11002167  0.017534618
#> 28 -1.3808804386  3.35380409  2.09958968  4.19543093  1.76742383 -0.942438344
#> 29 -0.0563143235  4.98697835  1.13108402  4.60936941  0.23296034 -2.539010113
#> 30  1.9940663837  6.11828142  1.59319705  3.46310696 -0.62447502 -4.696685239
#> 31  2.0528016227  5.09143998  1.81161297  3.24586340  0.76791207 -5.061487932
#> 32  0.3309812215  2.62883379  1.17707300  3.01521778  2.15704871 -3.776629459
#> 33  0.0901287475  1.22900967 -0.87992605  1.78361643  3.04001070 -3.078458053
#> 34  1.9634864343  0.70787435 -2.15780704  0.96211283  4.42554487 -2.267541682
#> 35  4.6502406986 -0.38836747 -1.28144466  1.98999393  3.67834410 -2.218191656
#> 36  4.8677012814  0.49545102 -0.95680146  3.06437722  2.45072347 -0.440377408
#> 37  2.4462820750  2.57629310  0.89962779  2.04983559  3.28564753  1.182623459
#> 38  1.4395456100  2.89281745  1.93426064  1.75217039  1.94475296  1.056158550
#> 39  0.8527435172  3.10113166  1.59385309  2.38049646 -1.17125366  0.811502359
#> 40  0.9153018929  1.01346274  1.49582629  2.91387975 -2.54862697  1.558590668
#> 41  0.6807110320 -2.20542378  0.30144729  1.93140184 -3.19356969  2.829998519
#> 42  1.3411287454 -2.99293928 -0.32522643  1.42595372 -1.36576634  4.173119128
#> 43  2.7471569763 -5.12676291 -1.18193739  2.34163322  1.00153081  2.655892535
#> 44  3.2073972086 -4.79244388 -2.13742445  2.45302873 -0.01643390  0.807078535
#> 45  2.2025540632 -0.92642680 -3.44261968  0.81433271  0.05989947  2.026963411
#> 46  0.6869465080  0.26978804 -2.40711212  1.24056660  2.26082261  0.687981519
#> 47  0.1235474699 -0.73145276  0.57781349  2.28441939  3.13763835 -1.766936562
#> 48 -0.4676608284 -1.87479071  1.17056896  1.03828016  2.66841322 -1.313469252
#> 49 -1.7729713435 -1.72300410 -0.17979076  0.89659521  0.70878939 -0.004328926
#> 50 -1.1214774776 -0.86789571 -1.16739105 -0.49069913 -0.62527177  1.980448424
#>         state42     state43     state44    state45      state46    state47
#> 1  -0.351777920 -4.56814581  3.04648907 -0.4211408  1.059557292  5.4181592
#> 2   0.244890511 -4.04979249  1.83550839  1.7539167  0.003282320  3.4748360
#> 3   0.315712897 -2.52942846  1.42876634  2.7730746 -0.313957469  2.3491749
#> 4   2.459408548 -2.04096634  1.73611720  1.5748292  3.544402180  3.8723813
#> 5   1.569441194 -0.21817730  2.00469573  2.1194596  5.445130853  6.7075089
#> 6   1.013871952  0.65566532  4.83985675  1.0017332  4.699647896  7.8878139
#> 7  -0.099645811 -0.28814769  6.50673170 -1.5080354  5.386233619  7.2099180
#> 8  -3.200805518 -0.30460321  5.51093969 -0.6547817  5.672172336  6.1358660
#> 9  -4.745826927  0.82584744  3.73002297 -0.5094307  3.803558229  6.1417139
#> 10 -4.240711730  0.89689431  3.11492284 -1.6900248  1.291717691  4.8836816
#> 11 -3.000125940 -1.14951112  3.04824663 -2.2218199  0.600154948  4.0444501
#> 12 -4.032785364 -2.64765713  3.16940749 -3.9726650  0.707819370  3.9117566
#> 13 -3.399212813 -0.97739578  3.36313925 -2.4802023  2.094840752  4.5558741
#> 14 -1.382730683 -0.52350278  1.30140463 -1.4167086  2.972634424  4.5890166
#> 15  0.316916144 -3.86933603  0.44425339 -3.7473886  0.219834848  4.6428554
#> 16  0.530191198 -5.59863855  1.52611653 -4.8397610 -1.645099884  4.4921440
#> 17  0.167483611 -2.58649727  1.80800346 -3.4698051 -1.337109930  4.0920683
#> 18 -1.158640583 -2.76100379  3.54082877 -0.2299180 -1.090620887  4.1645664
#> 19 -3.181841901 -3.40511300  5.24251178  0.5126285 -0.848635154  0.8834245
#> 20 -2.495029968 -1.72614810  4.21981180  0.7863053  0.261272068 -1.5161237
#> 21 -1.272231572 -2.39994604  2.53158118  1.6908596 -0.180306073 -2.1486787
#> 22 -0.752302400 -1.34459477  2.55645574  1.1276855 -1.203639207 -2.3653473
#> 23 -0.831794401  0.27343497  3.11518773  1.7240922 -0.607069197  1.1867073
#> 24 -1.025587756 -0.69580924  3.06377931  3.5710384 -0.882922101  4.0364179
#> 25  0.657071149 -1.17608629  1.59979595  5.1435674 -1.765383285  4.7532685
#> 26  2.112598493  0.92523425  1.47354229  2.3310981 -1.108535055  4.5814196
#> 27  1.406954487  2.44665215  4.17796739  0.2023241 -0.382796174  1.9486858
#> 28  0.407377176  1.17581323  4.64025630  1.3665162 -1.615359539  1.8579993
#> 29  1.538349291 -0.33991167  4.23919503 -0.3020642 -2.077456458  3.0125679
#> 30  1.626811470  0.89108059  4.78694407 -1.0754566  0.899298601  3.2873554
#> 31 -0.697558302  4.36506898  4.06553321  0.8781269  1.889481684  2.3711576
#> 32  0.455191121  4.52081504  2.84166857  1.0175289  0.980047883  0.1703324
#> 33  2.549046871  1.95947012  2.44067773  0.7102716  0.514463172 -1.5841292
#> 34  3.209900630  1.89354894  4.20747436 -0.2796556 -0.579748955 -1.5606650
#> 35  3.471311742  2.88457502  3.77534286 -0.4822475  0.033482510 -0.6188072
#> 36  3.146115251  3.07573205  1.29724158  1.0348198  2.355274562 -2.0573236
#> 37  1.437375254  3.31806356  0.65724425  1.0761097  4.186849539 -3.5596194
#> 38 -0.007064239  2.18545946  1.28292647  0.8329594  3.402588706 -4.3573257
#> 39 -0.450028198 -0.08431309  1.89880132  0.3857810  3.201287865 -5.5007733
#> 40 -2.708265513 -0.16113260  0.37782494  1.2830166  1.925358639 -3.6184689
#> 41 -3.458990655  2.03395535  0.68629702  1.9148735  0.004145527 -1.6409836
#> 42 -1.439201435  2.95017438  2.85477903  0.2087331 -0.493380469 -1.0343092
#> 43  0.714695752  1.84338415  1.67840804 -1.2823890 -1.097964899 -0.2096648
#> 44  0.776148153  2.13075605  1.03027150 -0.8539358 -1.258689661 -0.7878885
#> 45  0.025541581  1.85322072  2.64945297 -0.4129255 -2.028245024 -1.0231434
#> 46 -1.669413513 -0.68337028  2.50054171  0.5492587 -2.000040678 -0.5622784
#> 47 -2.611372942 -0.86930586  3.83326667  3.2242132 -0.353821202 -0.6253277
#> 48 -1.131328003  0.98895819  5.11982815  3.4571122 -0.824796963  0.4896308
#> 49 -0.273510944  1.79820858  2.84754007  0.3430256 -1.414419697  0.6304989
#> 50 -0.097294340  1.57036800 -0.03110999 -1.2433745 -1.034062310 -0.3153126
#>       state48    state49     state50      prob action      outcome      signal
#> 1   2.7814463  2.9717115  3.61515913 0.5424591      0   7.59401524 -3.70377972
#> 2   2.2010167  3.4472916  2.92358592 0.6122861      0  13.75576206 -2.51766069
#> 3   0.1422593  2.9018909  2.55004270 0.6652956      1  14.52578863 -2.76789547
#> 4  -1.2446175  0.9876205  3.55408077 0.4098444      0  17.11406140 -2.93671537
#> 5  -1.2116368  2.0353355  3.16349774 0.5863083      1   5.19408486 -4.11616813
#> 6  -0.1305774  2.3334674  4.03904294 0.3012961      0   2.51426511 -4.52047787
#> 7   0.1210236  1.8032632  6.03426165 0.4258415      0  -5.59982021 -3.89686884
#> 8   2.2357800  2.5127247  5.19420020 0.4479516      0  -5.95967340 -3.87676803
#> 9   1.4762149  3.4928563  2.52888199 0.5335589      0  -1.03689595 -0.48280592
#> 10 -2.1475666  4.0702519 -1.13141601 0.5529365      1   3.59760845  2.32126025
#> 11 -3.1590934  3.8950833 -3.70334822 0.3077989      1   7.25969524  2.62440428
#> 12 -3.4332864  2.6107949 -4.31405412 0.2751794      0  -1.32448263  3.47510578
#> 13 -4.7832292  2.9669554 -4.22919714 0.5489459      1   4.97238413  6.15080573
#> 14 -6.2117621  3.4343758 -3.33152141 0.3385893      1   9.44953808  7.38242438
#> 15 -5.4732177  2.2942421 -1.08162957 0.3115221      0   1.15226341  6.01355583
#> 16 -2.3028615  3.5773738  0.65682694 0.5290952      0  -4.97601313  3.58019075
#> 17 -0.9811068  4.9139180  0.80845540 0.5123629      1  -1.52740496 -0.51644022
#> 18 -1.0449212  4.1976064  0.28573040 0.2986480      0  -1.53468649 -3.37049940
#> 19 -0.8284510  2.1798076  0.58379403 0.5126030      1   1.22157796 -3.38587114
#> 20 -0.5754439  0.8502093  0.30734380 0.2238221      1  -5.73474984 -4.18030038
#> 21  0.1252142 -0.4764977 -0.46500019 0.1886206      0  -4.87950939 -7.36225318
#> 22 -0.5435165 -1.1428441 -0.48360363 0.3612481      0  -8.80356382 -7.89063872
#> 23 -0.1250371 -1.9220577 -2.07808670 0.4546758      1  -8.50508713 -4.06231923
#> 24  2.2747574 -4.3424038 -2.09279938 0.3167025      0  11.23046763 -1.37228734
#> 25  3.7312151 -3.3142142 -0.14449116 0.5764233      0  10.24785417 -0.46838655
#> 26  3.5258000  0.8178550 -1.33711540 0.6017551      1  10.27654158  1.05368620
#> 27  3.0540808  1.6061764 -1.18364728 0.4155726      0  17.05141597  3.29716425
#> 28  1.7819699  2.1518164  0.42349762 0.6548829      1  20.42526262  3.52561840
#> 29 -0.4123771  3.9336152 -0.18216256 0.3818307      0  13.31162880  2.14923974
#> 30 -1.7873917  2.7748499 -0.70156254 0.6412939      1  15.13774186  1.93653066
#> 31 -3.2562407  1.1585919 -1.27747674 0.4299921      1  17.92168154 -1.01922085
#> 32 -4.1538814  1.5992714 -1.70064245 0.4364317      0  19.17090332 -3.55770814
#> 33 -2.3601268  2.5202172 -0.83412948 0.6459349      1  13.46511293 -4.14260620
#> 34  1.1340646  1.9454186  1.17623777 0.3801163      0  14.63231045 -3.96743884
#> 35  1.0984966  1.9570957  2.19340463 0.6362238      1  11.74495724 -3.14716985
#> 36  0.3965931  1.4671798  1.99613857 0.4294439      0  16.42122110 -1.12761867
#> 37  2.2052753  0.3402847  0.87165622 0.6795880      1  14.26894434  0.07003118
#> 38  3.0821836  1.0563487  1.66219521 0.3121598      1  -0.01193276 -4.37360443
#> 39  1.4321063  0.5800170  3.18731066 0.2055455      0  -6.91909625 -6.70953210
#> 40 -1.4177360 -1.6464688  3.98436954 0.3948548      0 -10.07551513 -6.17516772
#> 41 -3.2796819 -1.8841515  4.02039869 0.3633959      0 -14.13490005 -6.92092681
#> 42 -4.9105220 -1.2146749  2.32758996 0.4031208      1 -13.73826955 -4.80601419
#> 43 -5.6554870 -0.2688177  0.22052313 0.2030294      0  -6.51493366 -1.91639165
#> 44 -2.7969254  0.7682163 -0.42622933 0.4425417      1  -7.79139275 -2.54689985
#> 45 -1.2186713  2.2127407  0.04691567 0.3257104      1   7.07195296 -3.73499028
#> 46 -3.4832489  3.5689068 -0.53313657 0.3539197      0  15.87560683 -3.48153456
#> 47 -3.7444709  3.3446687 -0.44270562 0.6098073      0  23.16299746 -3.38396114
#> 48 -3.4797733  4.4375205 -1.16410880 0.5836080      1  10.71064587 -2.32907999
#> 49 -2.2054061  3.7005340 -3.14843716 0.2603619      0   2.83431783 -1.25434197
#> 50 -0.2799870  1.2364228 -4.08541296 0.3591701      0 -11.55580386 -0.74823449
#>          noise
#> 1   1.32870598
#> 2   0.79093081
#> 3  -1.72237990
#> 4   0.02567727
#> 5  -1.82121276
#> 6  -2.81939488
#> 7   0.21169205
#> 8  -2.47251700
#> 9  -4.65544060
#> 10 -2.75370183
#> 11  0.70382769
#> 12 -1.15567600
#> 13 -2.71227524
#> 14 -3.69359470
#> 15 -2.14905520
#> 16 -5.99456201
#> 17 -2.51211020
#> 18 -6.19766766
#> 19  1.61126981
#> 20  3.59816760
#> 21  5.20677804
#> 22  2.59473947
#> 23 -1.74491206
#> 24  5.02004322
#> 25  2.27517687
#> 26 -0.46267131
#> 27  1.94600337
#> 28  3.19419950
#> 29  1.17690441
#> 30 -0.08151885
#> 31  0.54972610
#> 32 -0.99037845
#> 33 -0.09858724
#> 34  0.35058880
#> 35 -1.08546230
#> 36 -1.96001769
#> 37 -4.55047630
#> 38 -2.25255784
#> 39  0.50150267
#> 40 -1.84005495
#> 41 -2.63356068
#> 42 -1.05770124
#> 43  2.28265372
#> 44 -0.60027281
#> 45  2.78169644
#> 46  4.68976604
#> 47  9.93687313
#> 48  3.24039234
#> 49  3.60994197
#> 50  2.64964822