generate_individual

simulates data for an individual participant

Usage

generate_individual(
  id = 1,
  T,
  P,
  sigma_residual,
  sigma_randint,
  main_rand = 0.5,
  rho,
  beta_logit,
  model,
  beta,
  theta1
)

Arguments

id

the assigned id for this specific individual

T

the number of observations for this individual

P

the number of measurements or predictors

sigma_residual

the noise added to the moderator formula when generate pseudo outcome. This noise is generated using ARIMA

sigma_randint

the noise added to the moderator formula so the intercept is random across participants when generate pseudo outcome.

main_rand

the noise added to the control formula when generate pseudo outcome. This noise is generated using ARIMA

rho

this value is used to simulate state or predictors and random noises. We also use ARIMA when generating predictors, and rho defines how much of the past values influence the current value in an autoregressive model

beta_logit

the true value of beta when simulate the probability of assigning treatment \(p(A_t = 1|H_t)\)

model

the true model for moderator when generate pseudo outcomes

beta

the true coefficients for moderator

theta1

the coefficients for the working model of control variables. For now, we just simply assume the working model contains the all predictors with coefficients equal to theta1

Value

A data frame containing the generated states \(S_t\), action \(A_t\), outcomes \(Y_{t+1}\), treatment effects, and noise components for a single participant. The outcome \(Y_{t + 1}\) is generated following the typical WCLS setting i.e. $$ Y_{i,t+1} \sim g_t(H_t) \theta_1 + (A_t - p_t(1|S_t)) (f(S_t)^T \beta + \sigma_{i,t,residual} + \sigma_{i,randint}) + \sigma_{i,t,rand} $$ where i represents the ith participant and t is the tth observation for this participant. This function simply assume the working model \(g_t(H_t)\) is \((H_{1,t} + H_{2,t} + ... + H_{P,t})\theta_1\)

Examples

generate_individual(id = 1, T = 50, P = 50, sigma_residual = 1.5, sigma_randint = 1.5, main_rand = 2, rho = 0.7,
beta_logit = c(-1, 1.6 * rep(1/50, 50)), model = ~ state1 + state2 + state3 + state4,
beta = matrix(c(-0.2, 0.8, 0.3, 0.7, 0.3),ncol = 1),
theta1 = 0.8
)
#>    id decision_point      state1      state2      state3     state4     state5
#> 1   1              1  3.30474256  1.39582649  0.83095981 -1.0087947  1.0949433
#> 2   1              2  4.46374619 -0.60925809  3.16733676 -0.7964939 -0.5416294
#> 3   1              3  2.18406962 -1.32510412  2.71116494 -1.1241022 -0.3914890
#> 4   1              4  1.19388102 -2.18601017  2.59061100  0.1468918 -1.5270347
#> 5   1              5  0.63203664 -3.80420829  3.14394353  1.2309099 -1.0763520
#> 6   1              6  1.21804110 -2.38140194  1.86070453 -0.9963578  0.6832645
#> 7   1              7  1.24641487 -1.33408265  0.99277438 -0.9337048 -1.5235716
#> 8   1              8  0.17389692 -0.79006113  1.26795360  1.6081644 -3.2478918
#> 9   1              9 -0.04237168  0.70787715  0.90627435  0.8228109 -1.7275132
#> 10  1             10  1.16682360  0.59607884 -0.66320074  0.7906015 -0.8199790
#> 11  1             11  1.68393076 -0.72185226 -1.96888519  0.2916819 -1.0792191
#> 12  1             12 -2.13666709 -0.74328470 -1.83828487 -0.6862238 -1.4370575
#> 13  1             13 -6.78763071  0.74748548 -2.64238839 -0.8291165 -0.6663264
#> 14  1             14 -7.76391325  1.12862557 -4.25511807 -1.2221361  0.1589733
#> 15  1             15 -6.54058523  0.12626599 -4.81680332  0.4538461  1.5037203
#> 16  1             16 -6.80336307  1.37562983 -2.96626427 -0.6363973  2.8323302
#> 17  1             17 -4.70209660  1.98204983 -2.41860799 -3.3774160  0.6039777
#> 18  1             18 -0.85589496 -0.24776305 -3.46949193 -3.6022575  0.3507519
#> 19  1             19 -0.25770504 -0.74525425 -2.27296209 -3.9512374  1.9998055
#> 20  1             20  0.54055049 -2.42059477 -2.78060212 -4.5546672  2.7623294
#> 21  1             21  2.22628362 -2.16388891 -3.50930480 -4.1552970  3.0737050
#> 22  1             22  2.70312481 -0.05101425 -2.53714171 -3.9047930  0.6501806
#> 23  1             23  2.50663399  1.29029901 -2.37390048 -3.7969008 -1.9691769
#> 24  1             24  0.80472503  3.14371330 -1.96580882 -5.8361409 -0.5380812
#> 25  1             25 -1.12219304  4.72228318  0.08537030 -6.6296272  1.3787290
#> 26  1             26 -2.97261487  5.00925139  2.31695795 -5.1130609  2.7774801
#> 27  1             27 -2.13435730  3.83233976  0.94086971 -5.6722802  2.4225198
#> 28  1             28 -0.93046484  1.20442800 -0.64389814 -6.0545305 -0.4591951
#> 29  1             29 -2.01846496 -1.22448873 -1.44893312 -5.5890647 -3.1263858
#> 30  1             30 -0.48872627  0.18705743 -0.02459456 -5.9153611 -3.3744281
#> 31  1             31  2.57540421  1.63140548  2.22025731 -6.9175338 -1.9393718
#> 32  1             32  2.63851170  2.02424935  1.58194767 -7.8509857 -2.7769991
#> 33  1             33  0.25351463  1.88781875  2.22468976 -7.2100684 -3.3732958
#> 34  1             34 -1.51110241  1.13636468  2.90523442 -4.4552999 -1.6644338
#> 35  1             35 -0.48216146  1.64546240  2.20146424 -2.1819092 -0.7100386
#> 36  1             36  1.22715199  2.58652933  2.81476078 -1.6339488  0.3349980
#> 37  1             37 -0.18294523  3.17027682  3.72450021 -1.3538416  1.6444968
#> 38  1             38 -3.03829642  1.93841629  2.13743674  0.2875488  1.5446357
#> 39  1             39 -4.89230179  0.62194048 -0.28939303  1.9567785  3.6389864
#> 40  1             40 -4.84474141 -1.61556804 -2.98965584  3.1788914  5.6338189
#> 41  1             41 -2.83731894 -4.30330463 -3.38355884  2.6472676  4.1815508
#> 42  1             42 -1.79211750 -5.33847212 -0.99263438  2.0372456  4.2932853
#> 43  1             43 -2.06104944 -4.69939489  1.74182764  1.8142712  5.7450422
#> 44  1             44 -0.41384882 -3.41433779  3.56967292  1.5878285  5.5305946
#> 45  1             45  0.06572610 -3.63537213  3.24074048  1.7330988  5.6188879
#> 46  1             46 -1.55094331 -3.40146278  2.46882787  0.1342276  4.2710222
#> 47  1             47 -1.71858668 -2.18011634  3.01163931 -0.2178738  2.6113993
#> 48  1             48 -0.79200591 -3.30385407  4.48177992  0.3656345  3.1546285
#> 49  1             49 -0.10537249 -4.63157337  4.37408684  2.5442329  3.4351450
#> 50  1             50 -0.13863148 -2.58940642  3.21946893  2.4629368  3.1703428
#>        state6      state7      state8     state9     state10      state11
#> 1  -0.5096393 -0.04104464 -2.21289058 -5.5960593  1.54405473  2.874893388
#> 2   1.9959382  0.23667198  1.32276627 -4.2081283  1.39997189  3.067384571
#> 3   5.4425988 -1.95587258  3.46267399 -2.7088307  0.55408409  3.860138629
#> 4   6.6513475 -3.28281427  2.64305848 -2.8209598  0.24153973  2.227182166
#> 5   6.4121346 -1.57621345  1.32925173 -3.3733668 -0.60134424 -0.833103096
#> 6   4.6034009 -0.65458204  1.48311085 -1.1895766 -2.31933375 -0.883089589
#> 7   3.9074815  0.18863954  3.12649093 -0.2487087 -3.15090071  0.586342981
#> 8   2.3347583  1.53237131  2.27014086 -1.8965502 -2.31444902  1.637106082
#> 9   1.3120348  1.97007203  1.62136515 -3.5070446 -1.09921743  0.555399616
#> 10  1.5389144  1.11918069  0.63511602 -2.6295285 -0.20571064  0.213154079
#> 11  2.6202525 -0.59217633 -0.43040250 -0.8854214 -1.76384727  0.440830120
#> 12  5.3187989 -0.17816512  0.35471991 -0.5294868 -4.15370792 -0.977208123
#> 13  6.8608702  0.17229379  0.90257227 -0.5243378 -1.78810661  0.622744479
#> 14  6.4393609 -1.29496040  1.54110141 -1.0981022  0.50414160  1.941149226
#> 15  4.6064150 -1.99848174  0.58936970 -2.6961329 -0.50849191  0.956825730
#> 16  3.1432224 -1.31753472 -0.25585638 -5.1117740 -0.82065698  0.340388592
#> 17  2.6248250  0.82079621  1.21650072 -5.7981737 -1.02029478  0.028087506
#> 18  3.0348730  1.16158261  1.71616049 -4.9838404 -1.38510874  0.724060463
#> 19  3.1490994 -1.20445762  0.54697908 -4.9295155 -0.41002214  2.469388403
#> 20  0.3179774 -2.17415709  0.05128132 -6.0230749 -0.03599422  5.187263768
#> 21 -0.8901896 -0.57843026  1.09177416 -7.9440514 -1.45000545  5.099766457
#> 22 -0.3639382  0.82768013  2.14413212 -9.4718185 -1.67974744  1.070645791
#> 23 -2.3680667  1.12382271  0.09016969 -8.5630815 -2.04465808 -1.516885740
#> 24 -4.2279935  0.29507092 -3.88342562 -6.0712041 -2.22643055 -1.386641065
#> 25 -4.1822612  1.52899662 -6.63821638 -3.4630709  0.28766021 -0.007626383
#> 26 -4.0615365  3.10826432 -6.16526660 -1.2147938  2.96102975 -0.083153231
#> 27 -1.8622404  1.60353802 -3.96587038 -0.6000473  4.45978130 -0.480935091
#> 28 -0.5032326  1.04297801 -2.01770622 -0.6904331  4.84950700  1.180172681
#> 29 -1.3109204  1.37554815  0.10077285 -0.6201927  4.28213369  2.144105536
#> 30 -0.1532035  1.10102758  0.06093003 -1.6517662  1.93665132  2.030674417
#> 31 -0.2030137  0.63087710 -2.74916191 -1.5155700 -1.95547225  1.734961933
#> 32 -2.1207726  0.88516753 -1.99650224 -1.0806284 -3.03016353  1.575238550
#> 33 -2.6610174  1.35355992  0.12716630 -1.2056502 -2.13171441  1.116277920
#> 34 -2.6483939 -0.27628524 -0.82590035 -0.1491792 -1.94187270 -0.101052754
#> 35 -3.1339061 -1.50204725  0.05357491 -0.1866078 -1.55817145 -0.596564427
#> 36 -2.7914301 -0.79663212  2.81246495 -0.4660519 -1.32339530 -0.603387351
#> 37 -0.3458003 -2.90963809  4.28013660 -2.3097394 -2.36718754  0.990911285
#> 38  1.0096796 -3.98072607  2.92581521 -3.1695263 -3.06887936  1.474717418
#> 39  0.4141345 -4.40063340  0.83770466 -1.6234732 -1.33314231  0.602771717
#> 40  1.7520814 -5.07034306  1.21815100 -0.4671042 -0.60411414  0.736700992
#> 41  3.1287660 -3.05662232  1.96135256  2.9721639 -0.68385486  2.067720886
#> 42  2.9141077 -2.50003527  1.20681306  6.1914730  0.60364683  2.861520534
#> 43  3.1716583 -3.84278493  1.02309003  6.4392122  1.90480399  1.938990369
#> 44  3.2301581 -4.70512180  2.75845444  4.6900061  1.15303235  0.271900612
#> 45  3.1308663 -3.56597545  3.98445155  2.6326712 -0.38861825 -0.899530428
#> 46  2.4861124 -3.91274565  1.59180016  3.6148534 -1.98847985 -0.416705711
#> 47  2.1805349 -4.47004515 -2.08653587  3.3539145 -1.71962611  0.541847589
#> 48  3.2235176 -4.68390929 -3.24585690  1.5864816  0.17899013 -0.313377492
#> 49  3.4374375 -5.24003303 -2.71001994  1.3704421  1.21164736  1.185290374
#> 50  4.0125244 -2.64465092 -1.03117191  3.0246744  3.00698170  3.522863628
#>        state12    state13     state14    state15      state16     state17
#> 1   3.84366587 -4.0874907  0.42652264  1.3332276 -2.140503647 -0.51406660
#> 2   1.86235037 -3.5912788  0.76837200 -0.1793158 -3.657232344  1.65355932
#> 3   1.45751633 -3.4096340  1.36366036 -1.2377702 -4.510773785  2.44935990
#> 4   0.74234667 -2.0199894  1.58395004  0.1103516 -4.430485669  1.04480536
#> 5   0.46983005 -0.8181749  1.39721063  1.9207165 -2.489800247  1.72098083
#> 6  -0.47146491 -0.7522080  1.67556443  2.7856858 -0.900503488  1.98588067
#> 7  -2.59762558  0.1500178 -1.71095379  3.0293483 -0.008003627  0.33646719
#> 8  -3.86376241 -1.4609653 -2.67386934  1.5441688  0.496343415  2.29381756
#> 9  -1.81975523 -2.1470079  0.07821581  0.8167944  0.019451327  3.68290066
#> 10 -0.87376441 -1.4673570  2.79907254  1.1822080 -0.731332063  2.35479011
#> 11 -0.83546335 -0.7768641  2.54963406  1.7536656 -0.212523864  1.47889254
#> 12 -0.03943898  1.0192085 -0.32982016  1.0120364 -0.403448530  0.04804143
#> 13  0.70586189  2.7199207  0.24601956 -0.9479394 -1.514557526 -0.73061663
#> 14  3.23268878  4.2797501  2.20186704  0.1082169 -2.222998409 -0.91607571
#> 15  3.91750887  4.3333968  2.44733285  0.8238314 -2.418109996 -0.63292744
#> 16  3.31640335  0.2091700  3.09419821  0.5766916 -1.407538474 -0.95122593
#> 17  3.20194387 -3.9126471  4.16069722  0.5128026 -0.188281543 -2.35787066
#> 18  1.50771488 -3.0615583  3.95298278  1.5433887  0.560776276 -1.93998250
#> 19  0.21749343 -1.2895826  2.34819683  2.9816440  2.279996269 -0.87749654
#> 20 -0.91045285 -3.1588523 -0.04411140  2.7945564  2.263427643 -1.61827065
#> 21 -0.19253057 -5.4029836 -1.68437263  1.7916631  1.806007696 -1.86914982
#> 22  2.33716229 -4.8622944 -2.64736473  0.1143336  4.307351205 -0.85796833
#> 23  2.26703235 -4.7590394 -1.89088636  0.2497401  4.928152655  0.50744146
#> 24  0.95231866 -2.8270195  0.93107553 -0.7143800  4.154712505 -0.90668241
#> 25 -1.22293244 -0.2779811  2.72992689 -1.8628478  0.900635591 -3.27898096
#> 26 -3.21707700  1.0212526  3.03324376 -1.3831194 -0.881870716 -2.27512560
#> 27 -2.77804375  0.9028193  2.38094868  0.6042145 -1.894371052 -2.67620452
#> 28 -2.09860088 -0.4088592  2.76024792  0.8453157 -1.207110972 -4.04252891
#> 29 -0.31498803  2.0234411  3.51661733  0.4048820  2.625732156 -2.69128972
#> 30  1.00266190  3.8285835  3.40062918  2.2178327  2.547275373  0.26801186
#> 31  3.24433615  1.8091729  3.20464183  0.1960720  2.490270422  0.91912728
#> 32  3.94872400 -0.4439472  3.13495167 -2.0536803  4.040422775 -1.54811942
#> 33  0.18763015 -0.0268346  1.39618266 -2.4859286  2.475004031 -1.62083447
#> 34 -1.28285710  1.3378281 -0.66750027 -3.6250175  1.191958814 -0.62573512
#> 35 -0.10203685  1.6876561 -0.98335298 -5.8462827  1.146368896 -0.81703411
#> 36  0.28976502  3.4382623 -1.08908538 -5.9350236 -0.504142265 -1.24755090
#> 37 -1.45170237  5.2401531 -1.96215341 -3.2689005 -2.299221343 -2.79361408
#> 38 -2.59812617  3.3946924 -1.00357197 -2.2109180 -2.626329719 -1.28072357
#> 39 -2.78483706  1.3745172 -0.50972602 -2.2880898 -1.365787592  0.89507447
#> 40 -2.42389353  1.8139489 -2.16400298  0.0358720 -1.257366980  0.46099580
#> 41  0.86844882  1.2837390 -1.53264425  1.8049549 -0.853767139 -2.94134014
#> 42  2.92878348  0.6348487  0.22903145  0.0438026 -0.277940976 -4.89239739
#> 43  1.56913983 -0.1645747 -0.63874170 -1.8608273  1.005028539 -2.26464619
#> 44 -0.25585837 -0.9651058 -3.83826066 -1.7137553  2.626279805 -2.20791143
#> 45  0.02875915 -1.8095165 -3.40585389 -1.3864021  2.110935051 -1.63791772
#> 46  1.17151695 -2.7057023 -2.39437118 -1.6627686  1.746455374 -0.27057584
#> 47  1.21652448 -1.6998787 -2.14983085 -1.1352167  0.865859516 -3.74846740
#> 48  0.93376496 -1.5147252 -2.23909015 -1.1825147 -0.661058351 -5.53354824
#> 49  1.44175249 -1.1687796 -1.16823411 -1.0634269 -0.831860836 -4.01204791
#> 50  0.47774524 -1.8329151 -0.02149802 -0.6374945 -2.319070635 -3.96037330
#>        state18     state19    state20       state21     state22    state23
#> 1   4.31955824 -4.23277425  1.4635008  0.5508084944  0.99575452 -1.9073614
#> 2   3.18645362 -2.17507476  0.9038328  0.9981641575  3.65622800 -0.7492393
#> 3   2.14565268 -0.77375091  1.5020016  0.7713510525  6.50210153  2.0394594
#> 4   1.73939301 -2.27164654  0.5285541  0.2764668797  7.01351976  4.5648448
#> 5   3.20658207 -3.68788005 -0.4614098 -1.5258162894  6.33139760  5.0604457
#> 6   3.85934250 -2.83003181 -1.3095985 -2.3311813274  6.52053065  4.1209019
#> 7   2.25594059 -1.69041647 -2.0140560 -1.1570481195  5.47785227  4.6325463
#> 8  -1.04660964 -2.58472076 -0.4797670 -1.8918702666  4.05512987  4.7570889
#> 9  -2.47058929 -2.37104255  0.9509246 -2.4825985389  3.64231473  3.2656947
#> 10 -2.49350881 -1.58045389  1.5910513  0.0846306230  4.37981826  2.8960896
#> 11 -1.41187248 -1.47551622  2.1712631  1.7787410465  4.19263293  2.1474254
#> 12 -0.44695292 -0.58321572  1.3628085  0.4685937940  2.54142940  2.5224037
#> 13 -0.78267022 -1.25245725  0.7828954  0.8145144837  2.18253625  2.7813063
#> 14  0.93468470 -3.48580541  2.3784442  0.9206348013  3.04947519  1.7331300
#> 15  2.43538771 -4.52141012  2.6257455 -1.2881196361  2.86728446  0.2385171
#> 16  1.65168334 -4.92123926  1.3422656 -0.9358000665  1.59969302 -0.4696007
#> 17  0.59186847 -4.69929844  1.4472812 -0.2315840911 -0.43434701  0.2800931
#> 18 -0.42398349 -4.43467663  1.5469640 -0.9164859970 -1.77575872 -0.0827905
#> 19 -1.73564723 -3.22701548  1.5193947 -0.6260294411 -1.62990245 -2.3141400
#> 20 -1.71985926 -1.84981632  1.5527412 -1.1943066942 -0.74546973 -1.4265201
#> 21  0.43670201 -3.77659559  0.2838260 -2.9990988301 -0.67240352  1.9922751
#> 22  1.50134992 -2.65870376 -0.7769179 -3.4147580577 -1.40949310  1.9748856
#> 23  0.72083716 -1.02147962 -1.8643219 -3.6229142339 -1.89890788 -0.9132583
#> 24  0.01751420 -1.66582354 -1.6899135 -3.8990803835 -2.70016006 -1.1978227
#> 25 -1.06796986  1.16592570 -0.1708408 -3.5544162430 -2.72491279  1.1964814
#> 26 -0.75342730  3.36341511  0.4911124 -3.0864364417 -0.31454649  2.3208694
#> 27 -0.91229372  4.24629440  1.6652964 -3.8996931152  2.24245333  1.4078690
#> 28 -0.77162226  1.09936326  2.2645741 -4.6418768923  3.29051573 -0.9758751
#> 29  0.31740290 -2.37995441  1.5507315 -3.0266407479  1.80288157 -2.4498161
#> 30  1.21556268 -1.41259447  1.9412944 -0.7034038184  1.90426109 -2.2110280
#> 31  3.22921647 -3.13404187  1.9200724  0.0004854344  4.12442840 -2.0064570
#> 32  4.38217092 -2.60154103  1.4519556  0.8674042770  3.41376179 -2.7952269
#> 33  4.14309998  1.13415533  0.8217286  1.7730086498  2.94168049 -2.2169884
#> 34  3.32263078  1.52178353 -0.1832122  0.0125444807  2.76331725 -1.0272952
#> 35  1.49377752  1.86787697 -2.0545501 -0.7046961571  1.19517620 -3.0545660
#> 36 -0.80818726  2.50281271 -3.5554361 -0.3302187071 -0.24321259 -2.9459427
#> 37 -1.24772390  2.60174115 -4.0393131 -1.1517249413  0.01754069 -0.6218753
#> 38 -0.85820314  1.99466529 -2.9426676 -1.2504155554  0.26087174  0.2996072
#> 39 -0.36485332  0.89455551  0.2535904 -0.3770974502 -0.63317170  1.1043911
#> 40  0.27573363  0.05728957  1.1000896  0.1590912370  0.12829067  1.6186820
#> 41  0.08163084 -1.17431586 -0.5026208 -0.9645167894  1.70959487  1.8564337
#> 42  0.28163219 -2.38201733 -2.1695013 -3.1261263074  1.51704596  1.8478612
#> 43  2.98584211 -2.47354709 -2.8152333 -1.5425704948  1.14653490  1.6126383
#> 44  3.01600242 -2.54885087 -1.4173115 -0.9592806650  2.04579875  1.2953754
#> 45  1.53766713 -2.35729291  0.6444218 -2.5841834221  1.33173400  0.4978471
#> 46  2.72079290 -0.27568891  0.9665202 -1.9548380710 -0.26780703 -0.4028156
#> 47  3.45633294  0.76097928  1.6664576 -2.8342121339 -1.26849110 -0.9964759
#> 48  1.01005072 -0.44269166  0.9140037 -2.8843028085 -1.43903805 -0.8030027
#> 49 -0.84456299 -1.95914032 -0.3507391 -2.5245242576 -0.58654755  0.1887245
#> 50 -0.47406921 -3.00289962 -1.3900489 -3.4507596684  0.30762802  0.5656980
#>       state24     state25     state26    state27    state28     state29
#> 1   2.6384330 -0.58955647  1.66562411 -5.4322909 -0.4606993 -0.16196320
#> 2   2.1720905 -0.46963026  1.81577863 -5.5104317  1.6702911 -0.42391488
#> 3   2.4496312  0.01902743  0.73193192 -5.0726204  3.0539410  0.55001873
#> 4   3.5382671 -0.50851621 -0.28914838 -4.8811399  5.8178333 -1.42326744
#> 5   2.7916206 -2.09373721 -0.47121865 -3.5145170  5.6270920 -2.50776967
#> 6   1.0657214 -2.44453664 -0.49800275 -2.4169833  3.4221202 -1.53324542
#> 7  -0.5323359 -2.10967534  1.20203197 -1.4535497  2.7203764  0.05159359
#> 8  -2.8255083 -0.22693770  3.54919725 -0.7051514  2.1060754 -0.02439621
#> 9  -2.7032632  1.28699655  3.11817055 -0.7764054  2.4882299 -0.45991509
#> 10 -2.0284585 -0.15130771  1.48811433  0.5964204  4.0846822  1.23366332
#> 11 -1.2401796 -2.30218281  0.82150469 -0.9616361  3.9034023  1.54902460
#> 12  0.8664867 -1.34644888 -0.84236165 -1.0943660  3.6441245  1.14426908
#> 13  2.6004706  1.23987758 -2.22208298  0.3199090  3.2372677  1.54965238
#> 14  5.0074420  2.64084561 -3.50081952  0.3415369  1.1790770 -0.26016090
#> 15  5.5510086  3.48146085 -4.53751232  0.7910296  1.5349677 -1.43175838
#> 16  4.2073944  2.09196111 -3.51072677  0.2587548  1.5652227 -0.53352828
#> 17  2.1702259  0.30086774 -1.22130068 -0.1055778  1.7436236  0.03800895
#> 18  0.5815270  0.79105412 -0.61036849 -1.5775999  2.2858549  0.73182399
#> 19  2.6290030  1.21115027 -0.04926106 -0.4965914  1.3574374  1.01623378
#> 20  4.9850898  1.11720367  3.34177972  0.7316138  1.5413671  0.90854015
#> 21  5.7256484  1.33048467  5.28006625 -2.2980722  0.5857314  1.71702519
#> 22  5.8180730  1.98502995  5.41780272 -3.2093502 -0.5247143  3.39162181
#> 23  4.9095020  2.12169833  4.19373710 -1.0834081 -1.1353564  2.82841820
#> 24  4.1399640  1.13578856  2.55780734  0.7013373 -1.1276813  2.18237664
#> 25  2.8545402  0.10631141  2.34538116  2.2193051 -0.4913122  2.54216412
#> 26  3.8007191 -0.34752149  1.99047964  2.1713492  0.1328112  2.19372327
#> 27  5.2418772 -0.05003386  3.31133845  0.3683092  0.7725708  2.65768557
#> 28  3.0966368 -0.45274705  4.56835234  0.6882830  0.4648361  2.40183353
#> 29  0.2952548 -1.95086253  4.09386940  2.5433417 -0.1491068  1.57306813
#> 30 -2.3407350 -3.31170170  2.79926138  3.4162984  1.2371139  1.07934993
#> 31 -3.2120723 -3.20537810 -0.06144647  4.2100716  2.3600665  1.83891459
#> 32 -2.7491959 -2.79026666 -3.16011156  4.5138449  1.3031228  1.55906614
#> 33 -1.8047723 -2.33665981 -4.59494585  4.3194784  1.6018685  0.68185946
#> 34 -1.7566191 -2.38721980 -2.15745488  5.6282479  2.5074693  1.44367354
#> 35 -1.9269383 -1.34022812  0.69010177  4.3061393  3.2951462  1.86587892
#> 36 -0.9906613 -0.64182352 -0.77163967  2.2391191  2.7476623  0.86338201
#> 37  0.2719671 -1.54331784 -4.15309304  3.5326918  0.7840625 -0.71951027
#> 38  2.7933486 -2.06641561 -6.40008005  3.7802169 -1.3270527 -1.23229673
#> 39  2.5509743 -3.01617470 -7.37331118  2.3628737 -0.9505827 -0.71206466
#> 40  1.0540678 -4.16052092 -5.96040926  1.1786407 -0.4476584  1.25195466
#> 41  0.5973515 -2.95949855 -3.36247500  1.4406249 -0.4378484  3.20641998
#> 42 -0.9457626 -0.15706561 -2.72086228  2.2020302  1.0964278  3.14607400
#> 43 -0.7950357 -0.06503873 -1.60816415  3.5505119  1.5738582  0.83125214
#> 44 -0.5542179 -1.58430797 -1.24187687  3.3112981  1.4179267  0.30969250
#> 45 -0.9038965 -2.98058163 -3.09671432  2.9200641  2.9668962  0.07456968
#> 46 -0.3455526 -0.73312442 -2.09726932  6.1575785  3.2875782 -1.74760713
#> 47 -0.2640489  1.77845735 -1.09577149  7.6499644  1.5639530 -0.90558064
#> 48 -0.5337468  0.41154021 -1.08852371  6.7689947  1.5931198  0.52585049
#> 49 -1.0291511  1.24750325  0.13290831  5.4926515  1.0422598 -0.29984289
#> 50 -0.7321507  2.30106137 -0.05064492  5.2539199 -0.7059065 -1.00283600
#>        state30    state31     state32      state33     state34      state35
#> 1  -1.14863286 -3.9200544 -2.11193589 -0.379403018  3.75182045  2.605997118
#> 2  -1.91359590 -2.2556153 -2.61148828  0.843397696  1.89442884  3.506029420
#> 3  -0.74405772 -0.9696220 -4.63134812 -0.538071187 -0.57549865  4.900494703
#> 4  -1.04965750 -1.8548091 -3.90187975  0.008423396 -0.09911065  6.049805116
#> 5  -1.35652826 -2.2954376 -1.29501017  1.276172588 -0.02146891  7.247196548
#> 6   0.10805979 -4.1547684 -0.27474794  0.469873949  0.46651623  7.002155857
#> 7   1.62573500 -5.4718732 -1.89293922  1.070208080  0.57694862  4.604732045
#> 8   2.96847619 -3.1414724 -2.07269209  1.578157933  0.67543193  2.113245966
#> 9   2.29849898 -1.9166918 -1.58593665  1.392205353  2.47851506  0.006306097
#> 10 -0.91938869 -0.4551553 -2.16955456  1.185762167  0.23852118 -1.259896973
#> 11 -2.92682418  1.2564218 -1.70950379  0.849824546 -4.97491434 -2.508907550
#> 12 -3.68960698  0.9636728  0.03904275  1.345521934 -5.02993598 -4.508355371
#> 13 -3.10834297 -0.9081299  0.68435910  2.461531784 -3.66862884 -5.447901594
#> 14 -4.48797448 -0.5915375  0.36519733  2.872271896 -4.52684139 -3.241962074
#> 15 -5.47656426  1.8104505  1.06662469  3.196839102 -2.69775677 -0.822398691
#> 16 -2.58684517  2.5996935  2.59034450  2.632548866 -1.18886033  1.271874735
#> 17 -1.14148071  3.1129371  2.45015583  2.414230087 -0.39581227  2.244438211
#> 18 -0.49865683  2.7372877 -0.17298501  3.464488152 -0.88859895  3.387154494
#> 19 -1.58418943  1.9454164 -1.62036247  3.313680616 -0.47176604  5.023151011
#> 20 -2.63400616  2.3706273 -1.74800426  2.283467922  0.76437039  6.145419942
#> 21 -2.05061442 -1.1543420 -1.99947121  1.998570859  1.93584575  5.880976341
#> 22 -2.36693637 -1.8448531 -1.09507832  3.382441337  5.65267895  5.103529572
#> 23 -1.96522430  1.2129316 -1.53848228  2.894189065  4.28343193  4.408759583
#> 24 -3.41746688  1.3507539 -3.41795281  2.031060728  0.76136673  2.626174728
#> 25 -7.10006790  0.7399754 -2.88528281  1.820239412 -1.01343588 -0.699084106
#> 26 -7.67438962  2.3838955 -4.10215691 -0.343762922 -2.82341023 -0.821871660
#> 27 -4.60779745  3.4674142 -5.18218728 -2.108317999 -4.67702067  0.136058939
#> 28 -2.04994659  1.9755453 -3.57417564 -0.892069860 -5.36716676 -0.438991767
#> 29 -1.60407907  3.0055384 -2.74257671  1.250859236 -3.83872676 -0.541001432
#> 30 -2.56362407  2.5316494 -2.34338393 -0.619353595 -2.70078034 -0.351310794
#> 31 -3.25761621  0.5602251 -2.41625161 -2.432002929 -2.45948902  1.105436163
#> 32 -4.12507841 -0.9664499 -2.02797604 -2.612034454 -3.27152229  1.159274457
#> 33 -5.60873663 -2.9848584  0.17964361 -3.016284663 -5.41403196 -0.531721156
#> 34 -5.34274816 -3.6245821 -0.03049655 -2.447232839 -5.19487817 -0.788950738
#> 35 -2.22628071 -3.2601078 -1.73420552 -3.432965732 -3.10082294  0.599244782
#> 36 -0.61545084 -1.9129378 -2.29991370 -4.597923289 -1.55246506  1.205616127
#> 37 -2.29785828  0.4258388 -1.04411957 -4.363157312 -1.06462208  2.051182728
#> 38 -3.61302123  1.6566653  0.13773904 -3.942359971 -2.35297004  3.592863721
#> 39 -1.74642831  3.5135182 -0.61705169 -2.559617041 -2.06493786  5.928531925
#> 40 -0.02415206  1.9509944 -0.82899839 -0.557699847 -1.33802381  6.209620188
#> 41 -0.11763346 -2.5380540 -1.46829350  0.309523785 -2.86518255  5.998655332
#> 42 -1.29071296 -4.8683171 -0.57489590 -0.302217786 -5.67721384  6.255523127
#> 43 -1.62950485 -4.3289586  3.13838005 -1.055632282 -6.46184288  3.744197866
#> 44 -0.17677186 -4.1434226  4.57784956 -1.000265752 -5.27680293  2.253247886
#> 45 -1.02615456 -6.6764862  3.33898898 -1.659467727 -5.47724294  2.613699858
#> 46 -2.39834986 -5.8833945  1.67825201 -2.859635475 -3.07550134  1.861278264
#> 47 -2.84648371 -3.0651399  0.86434667 -1.566413754 -1.06932950  1.444033651
#> 48 -2.07723673 -2.5724298  1.78584725  0.161218544 -3.41472898  1.937843676
#> 49 -1.33319595 -1.0666950  4.13597765 -0.318737432 -5.30906251  2.929327419
#> 50 -1.21332172 -0.5309813  3.29954272 -0.679683989 -4.40344530  2.374350609
#>        state36     state37     state38    state39     state40     state41
#> 1  -1.41870142 -0.94874274  1.58106439  1.0733569  1.81795725 -0.08905778
#> 2  -3.30294507 -2.26027098  0.50432812  2.2631491  0.60614678  0.63412103
#> 3  -5.20745198 -4.31047551 -1.19798644  2.1612591 -0.38245122  0.44468321
#> 4  -5.08437789 -5.97866849 -0.70583493  0.3079530  0.70844409  0.90308905
#> 5  -2.75442997 -5.91110436 -0.13993145 -1.3313041  1.11566852 -0.66781175
#> 6   1.88896455 -3.98860712 -1.28504080 -2.0430413  0.71090271 -0.80473493
#> 7   2.89765404 -2.90695222 -2.55938152 -3.5679208  0.92384442 -0.46797272
#> 8  -0.55089257 -1.91420747 -4.01456230 -4.6736158  1.16158234 -1.91384712
#> 9  -0.33345349  0.57720460 -3.27848650 -2.4529831  0.06737255 -3.55430959
#> 10  0.25772903 -0.34119517 -2.21855996 -0.2375286 -0.54367209 -3.13662234
#> 11  0.90272279 -1.00598632 -2.28518746 -0.0864534 -1.47219799 -1.37841628
#> 12  3.47930382 -1.59924304 -0.75549195 -0.6110135 -3.32809367  0.22745664
#> 13  5.76528307 -3.22738678 -0.47810718  0.1285232 -3.64272711  1.46336927
#> 14  7.46664465 -3.51717717  0.07536892  1.4627275 -3.45816038  3.11789478
#> 15  6.29397239 -5.26170795  1.23000013  1.8971613 -3.90432574  3.26235066
#> 16  3.87593298 -6.29589244  1.50280262  2.4680857 -4.85593273  2.43187015
#> 17  3.16266018 -6.29261034  1.73914964  2.0439800 -3.86424291  2.00834468
#> 18  2.06604184 -5.46644255  2.57870970  1.5662814 -1.31680997 -1.37376531
#> 19  0.75267732 -3.06663051  2.66352347  0.6849019  0.13717592 -3.21243549
#> 20  0.84041752 -0.54061460  0.66347946 -1.2321104  1.76979182 -1.39561508
#> 21  0.10682788  0.83720229 -0.77966548 -2.6882008  3.14657254  0.46081489
#> 22 -1.66199782 -0.34728478 -1.40697815 -3.2996398  3.09853341 -0.55127002
#> 23 -3.15038887  0.05926841 -0.81724077 -1.5254978  0.86159370 -2.61264027
#> 24 -3.56060376  1.84655132  0.26885327  1.5457917 -1.39656638 -2.54631165
#> 25 -2.48887594  0.72423815  0.75825878  1.0412542 -1.14898495 -3.46256585
#> 26 -0.01475289  1.37509851  2.07482454 -0.7558303 -1.63438523 -4.18859788
#> 27  0.87457488  3.16277697  1.95282321 -1.3050990 -2.00028029 -2.86079022
#> 28 -0.40787861  4.40115733 -0.36755593 -1.3042094 -2.45946988 -1.68896010
#> 29 -1.18761577  4.65386933 -2.29674318 -1.6757825 -3.05828983 -2.34878873
#> 30  0.06443751  1.52438574 -2.90333286 -3.7054069 -1.33238033 -2.75579321
#> 31  1.07845687  0.62566710 -2.31829955 -5.5199430  0.31742510 -1.39220376
#> 32  0.13411858  2.96453632 -1.53747533 -6.7546537 -0.35407602  1.24944772
#> 33 -0.11202967  1.85776435 -2.07617382 -6.0738119 -1.08083820  2.61362635
#> 34 -0.04730971 -0.26449900 -1.74316749 -5.9175640 -2.45008409  2.66363549
#> 35  0.16656753 -0.79746661 -2.24671760 -6.4481978 -4.83265141  2.28225060
#> 36  0.31390686 -0.82937536 -3.80538065 -5.3151432 -4.67902126  2.10126484
#> 37  1.17226863  1.86674471 -4.52675903 -3.8250743 -2.74281787  2.59745728
#> 38  2.28819549  2.26233446 -4.72725202 -1.8675875 -1.45399179  1.27078750
#> 39  1.03760594  0.38024140 -3.61142711 -0.8102354 -0.53182106 -0.06698400
#> 40 -1.50466563 -1.89862466 -4.87647859 -0.8178532  0.31296388  0.30665084
#> 41 -3.29693524 -3.50663180 -7.09600098 -1.4679481 -0.60937841  0.20326675
#> 42 -2.49149954 -2.27588596 -6.93292897 -3.6542573 -0.86383257 -0.01622319
#> 43  0.78132502 -0.60670743 -4.77991002 -4.4798555  0.76635337 -0.55156097
#> 44  1.49524289 -0.15784503 -2.44058296 -2.9855737  0.84582727 -2.74636940
#> 45 -0.34737841  0.09081876 -0.98146433 -2.6472733  0.15615406 -3.80331323
#> 46 -1.40259684  0.99018294 -1.49513297 -3.6337231  0.93575257 -2.15794850
#> 47 -2.54958069 -0.98751769 -2.38433351 -4.1269742  1.31093014 -0.51094135
#> 48 -3.33016591 -3.11965072 -1.79868994 -3.0810136 -0.30327366 -1.16685700
#> 49 -3.50472688 -4.26749686 -0.58264653 -0.7936831 -0.82413418 -0.13807437
#> 50 -1.79732280 -3.66693840  0.38191420 -0.6416347 -0.76546943  0.73934375
#>         state42      state43      state44    state45     state46     state47
#> 1   1.102797014 -5.470663929  4.253813220  0.7873873 -5.07547310  0.27115161
#> 2   1.332158332 -2.544789503  3.251290213 -1.2180792 -3.04821563  1.47391832
#> 3   2.746462897  1.758121673  1.015494853 -0.7220274 -1.06250171  4.40857910
#> 4   4.148581983  3.450106633  0.357615483  0.5370866 -2.45142007  5.15294207
#> 5   3.976677803  1.598493927 -1.093530143  1.7295812 -3.36205537  3.70776909
#> 6   3.669064542  1.346843319 -3.591814549  3.7880194 -0.66033169  1.54218474
#> 7   3.474290282  0.571653440 -3.323268598  3.3085512  1.23728954  0.56842133
#> 8   2.437867474 -0.184730635 -2.326803829  0.2504358  2.96815110  0.27360975
#> 9   1.722480931  0.621356336 -1.903409592 -0.5150308  4.31358673 -1.09285437
#> 10  0.756440795  0.914340987  0.865803418 -1.0775606  3.38998415 -1.54623231
#> 11 -1.072801823  0.846321381  0.863120037 -1.0092914  2.54890702 -0.81103820
#> 12 -1.582805614  1.229723913 -0.292690908  1.8323095  3.55924088  0.73720867
#> 13 -0.682396856  3.345974914 -0.688125460  2.6575849  4.78166937  1.15799492
#> 14  0.004850269  4.645046229 -1.128030731  2.1537841  4.19648682  0.14242304
#> 15  0.919145469  3.484265716 -1.108227686  2.6226106  2.55597636  0.03991825
#> 16  2.170024155  1.744755945 -3.573116230  3.1196949  3.88903436 -0.53831552
#> 17  3.142151439  1.350851531 -4.292136558  3.6157946  3.59335146 -1.53639516
#> 18  1.410951034  0.508044173 -2.736795062  1.5022281  1.18791114 -1.33114502
#> 19  0.839057826 -0.631055260  0.365708157 -2.8433604 -0.43495014  0.52309259
#> 20  2.747655522 -0.008750028  1.824586804 -2.3984093 -3.09493159  1.64400507
#> 21  3.883836711  2.092510669 -0.664309505  0.5005751 -4.24595468  0.32969829
#> 22  4.969376683  1.976571555 -1.062298762 -0.5391455 -2.18097274 -1.85776303
#> 23  2.850469525  0.722090905  2.828998761 -1.1278512 -0.88225537 -3.03805755
#> 24 -0.414164873  0.626269531  4.916969245  2.2010652 -1.59891352 -2.92296291
#> 25 -1.325833608  1.773939612  2.176411821  3.6225813 -0.54517027 -4.28316776
#> 26 -1.009629162  2.300351729  0.447658460  4.6386073  0.88741163 -5.58224211
#> 27  0.250979524  1.681639783 -0.862486259  5.4076031  1.46252866 -3.63534527
#> 28  1.328647587  1.427665422 -3.363866941  2.9859562  2.43901753 -0.59377387
#> 29  0.082199438  2.095644830 -4.312205199 -0.2292832  2.52891370  0.63179845
#> 30 -1.141133691  2.592316876 -4.292285687 -1.7425513  1.32918069  0.41513972
#> 31 -0.333802735  2.461499819 -4.349090203 -1.4222553  0.08886981 -1.26915236
#> 32 -1.755094819  0.819948860 -3.801934226  0.2652610  0.84580839 -1.35366090
#> 33 -4.715895948 -0.570158106 -3.622052623  1.8814415  1.05960336 -0.48844438
#> 34 -5.923242885 -0.344707571 -3.155466912  2.6224481  0.07939045 -1.56358290
#> 35 -3.543305894 -0.833893856 -1.227673914  4.1780901 -0.92378808 -0.67434543
#> 36 -0.902339671 -1.069842200 -0.129493757  3.5372555 -2.93073807  0.38806344
#> 37  0.256274657 -1.405454101  1.209582792  1.1528872 -2.43071317  0.47014250
#> 38  0.936465647 -1.832913496  2.050033390  0.2690494 -2.33960570  1.10115576
#> 39 -1.516657136 -3.435806468  0.381391200  0.7392205 -3.40357886  1.16478602
#> 40 -3.603157809 -4.465524073  0.002502771  4.0364994 -4.16553985  1.76719515
#> 41 -3.095497909 -3.121108693  0.357166213  5.3486960 -4.43574513  1.44500297
#> 42 -2.019837365 -2.348460504 -2.250802925  4.6675868 -4.92965636  0.84255399
#> 43 -0.966133599 -2.009134749 -3.191953203  5.4727140 -4.08175807  1.31711119
#> 44 -0.025921444 -1.808249981 -1.348200845  5.9646049 -1.66722503  1.70140545
#> 45  0.487408504 -2.380074278 -0.449256199  7.1116972 -0.43899168  3.02510645
#> 46  0.693013383 -3.027548228 -1.034699196  7.2104358 -0.90297747  3.26869118
#> 47  1.163404670 -2.886474937 -2.250236821  4.9465414  0.17651965  1.43706730
#> 48  0.228968519 -2.444837362 -3.028366670  2.6380427  3.63969494 -0.73915775
#> 49 -0.405838456 -2.473634501 -4.037532459  0.9590846  4.62549583 -0.81099749
#> 50 -1.436307128 -2.363116098 -3.866070232  0.9033066  3.08363867 -1.23589978
#>        state48     state49    state50      prob action     outcome      signal
#> 1   0.90715079  5.06595343 -0.4112710 0.5130522      1   2.4444602  3.14157544
#> 2  -3.68552886  4.72726289  0.2987712 0.3357172      0   9.1831096  5.16640710
#> 3  -6.38192332  2.07647427  2.5759835 0.6260589      0  20.7149488  2.71030926
#> 4  -4.13308471  0.64792291  2.5384166 0.6295046      0  14.9270015  1.95679700
#> 5  -0.98970577  0.65081100  2.8125650 0.6202825      1  10.5431187  1.73440026
#> 6   2.36264163  0.03724544  2.6941749 0.4160484      0  19.0631116  1.06359814
#> 7   4.21522106 -1.40336624  1.9170874 0.6165203      0  13.8704554  0.81173772
#> 8   3.79917200 -3.51258655  2.5120046 0.5002114      1  -5.0043864  1.07211603
#> 9   2.25115954 -4.39765765  0.3563221 0.2733237      1   0.5088018  0.85970112
#> 10 -0.37858787 -3.52396177 -2.6679962 0.2832558      1   3.1253024  0.68522245
#> 11 -0.86993768 -0.98498297 -2.5801886 0.2234101      0  -5.9968360 -0.36012614
#> 12 -1.85647320  0.02696788 -1.2434418 0.4322597      0  -2.8307508 -3.62498564
#> 13 -1.46539008  0.43584383  1.0180781 0.5666493      1   2.9113587 -7.50426575
#> 14 -0.02660357 -1.45742114  3.1136679 0.4178063      0  21.9023008 -9.41776642
#> 15 -1.40267935 -4.00166655  4.7475665 0.6278594      0  20.0297558 -8.63019688
#> 16 -2.46179810 -3.96419329  4.9753032 0.5539532      1  -0.5451070 -7.49730567
#> 17 -2.25647338 -3.63575533  3.0265722 0.2800664      0   2.3113961 -6.07331273
#> 18 -0.68515280 -2.29926486  2.4345397 0.4775828      1  -4.9340913 -4.46836648
#> 19  0.30061671 -2.41780832  2.1810399 0.2699271      1  -0.7869633 -3.40618497
#> 20 -0.30151506 -3.97123172  0.4341760 0.2791377      0   5.1191823 -3.80655967
#> 21 -0.06770083 -3.47741152 -0.7285208 0.4694128      0   4.9859497 -2.77124222
#> 22  0.10488216 -2.20975766 -1.3662820 0.4869718      0   3.5729573 -1.00024152
#> 23 -0.58908374 -2.93117320 -2.3436320 0.3785974      0  -4.0780176 -0.60840367
#> 24  0.68486161 -0.58694566 -3.8560765 0.2984696      0 -14.4785149 -1.74001441
#> 25 -0.89963641  3.13780622 -3.6529283 0.3009164      0 -15.1916787 -1.61019844
#> 26 -4.06728565  3.04098105 -1.8419547 0.3983250      0  -6.0453564 -0.98736418
#> 27 -5.12030231  1.92105802 -2.1109307 0.4838557      0  -2.5173795 -1.80085919
#> 28 -5.15749613 -0.34856259 -2.5550741 0.4039843      1 -11.1839509 -2.85013132
#> 29 -4.41391842 -4.79391247 -0.9687889 0.1650319      1 -14.0875250 -4.87309119
#> 30 -3.42181354 -8.24430909  0.2411030 0.1676138      0 -10.4782794 -2.32668830
#> 31 -0.78586831 -9.16290362  0.4744759 0.3527598      0 -12.3683059  1.82866500
#> 32  0.78273575 -7.74012308  0.3275657 0.2936489      0 -18.0642331  1.27015183
#> 33  1.24804959 -5.20234029 -1.0651451 0.2346073      0 -23.8469644 -0.03658035
#> 34  0.80893651 -3.92485832 -1.9267211 0.2066900      0 -28.6669397 -0.37089840
#> 35 -0.03030108 -3.08158250  0.8429333 0.2397873      1 -22.9897131  0.79436175
#> 36 -1.65530101 -1.74860805  2.7714746 0.1285869      0 -18.6979402  3.03782830
#> 37 -2.94341096 -0.98158137  1.3148209 0.3198898      1  -7.2285422  2.80572452
#> 38 -2.00908203 -0.87324653  1.1718953 0.1478458      0 -14.5167083 -0.46664190
#> 39 -2.08835335  0.25465516  2.7182749 0.3327635      0 -14.1817033 -3.54280086
#> 40 -4.29662963 -0.46903059  2.9183247 0.3331068      1 -22.6419882 -5.69955519
#> 41 -5.64034045 -3.05534870  2.8080554 0.1417366      1 -20.7651033 -5.33515745
#> 42 -5.96485162 -3.95705681  3.7296655 0.1298999      0 -20.0992842 -3.31890602
#> 43 -7.17979987 -3.91782286  2.4327911 0.3870571      0 -10.3474349 -1.49509730
#> 44 -4.88086050 -3.82959195  0.3941226 0.4659615      1  -2.6092729  1.41973919
#> 45 -0.58032448 -5.60596349  0.2856218 0.2050402      0 -10.2439860  1.55041721
#> 46  1.17221032 -5.27876036  0.1122797 0.4149209      0 -13.0391161 -0.69274569
#> 47  2.20559607 -2.75887120  1.1445106 0.4195669      0  -9.6370648 -0.18611887
#> 48  1.41521638 -0.34492835  1.7693218 0.3499047      1 -16.7007460  1.42217534
#> 49  0.83396410  1.08206189  2.5363778 0.2098301      0  -9.9303112  2.15136064
#> 50  3.45419784 -0.11545559  3.9488444 0.5064861      0  -0.6994594  1.90478218
#>          noise
#> 1  -0.39083825
#> 2   2.97853564
#> 3   9.52810331
#> 4   2.90651371
#> 5  -2.38415026
#> 6   2.98103433
#> 7   2.50079019
#> 8  -5.56135905
#> 9  -0.67032087
#> 10  0.84339089
#> 11  0.07028844
#> 12  2.41825103
#> 13 -0.54149749
#> 14  1.26213049
#> 15  1.53509700
#> 16 -2.61736990
#> 17 -0.78622543
#> 18 -0.35652012
#> 19  1.57461157
#> 20  2.77520238
#> 21  6.74763313
#> 22  4.38897877
#> 23  8.07928361
#> 24  6.36704953
#> 25  5.39723193
#> 26  3.87257705
#> 27 -1.77373898
#> 28  0.23704919
#> 29  5.51272253
#> 30  4.19758525
#> 31  3.45002155
#> 32  4.25193234
#> 33  5.70635820
#> 34  4.88124771
#> 35  5.25255413
#> 36  4.53045588
#> 37  9.72020968
#> 38  4.20423853
#> 39  2.03219800
#> 40 -1.48682966
#> 41  3.83738626
#> 42  2.01568999
#> 43  0.56636418
#> 44  0.04165796
#> 45 -1.04895708
#> 46 -4.73507561
#> 47 -1.60136159
#> 48 -2.13883821
#> 49 -1.33013830
#> 50 -0.38336138